### 2025-11-01 cdd3f17
Add Bharat Khandelwal as an author
Added Bharat Khandelwal as an author with ORCID and affiliation.
