### 2025-11-28 be462ee
Reduce paper.md word count to 474 words (target: ≤1000)
Co-authored-by: phanindra-max <51989959+phanindra-max@users.noreply.github.com>
