+++ paper.md
@@ -23,373 +23,57 @@ bibliography: paper.bib
 
 # Summary
 
-This paper presents LAiSER, a comprehensive artificial intelligence framework designed to extract, standardize, and align skill information from unstructured textual data with established skill taxonomies. The system addresses a critical challenge in workforce development and educational policy: the lack of standardized, machine-readable skill information that can facilitate communication between learners, educators, and employers. LAiSER employs a novel two-stage pipeline that combines advanced language models with semantic vector search to achieve high-precision skill extraction and taxonomy alignment. The framework supports multiple computational backends, including GPU-accelerated inference, cloud-based APIs, and CPU-only environments, making it accessible to diverse research and operational contexts. Initial deployments demonstrate the system's capability to process large-scale datasets while maintaining accuracy in skill identification and classification.
+LAiSER is an artificial intelligence framework for extracting, standardizing, and aligning skill information from unstructured text with established skill taxonomies. The system addresses the lack of standardized, machine-readable skill information needed to facilitate communication between learners, educators, and employers. LAiSER employs a two-stage pipeline combining large language models (LLMs) with semantic vector search using FAISS [@Johnson2019] for high-precision skill extraction and taxonomy alignment to standards like ESCO [@ESCO2020]. The framework supports multiple computational backends—GPU-accelerated inference, cloud APIs, and CPU-only environments—making it accessible across diverse research and operational contexts.
 
 # Statement of need
 
-The contemporary labor market is characterized by rapid technological change, evolving skill requirements, and a growing disconnect between educational curricula and industry needs [@Autor2013]. Traditional approaches to skill extraction rely heavily on manual annotation, expert judgment, or simple keyword matching—methods that are labor-intensive, error-prone, and fail to capture the semantic richness of skill descriptions [@Boselli2018]. LAiSER fills this gap by providing an accessible, flexible, and accurate framework for automated skill extraction and taxonomy alignment, enabling researchers, educators, and workforce development professionals to analyze skill demands at scale.
+The contemporary labor market is characterized by rapid technological change, evolving skill requirements, and growing disconnect between educational curricula and industry needs [@Autor2013]. Traditional skill extraction approaches rely on manual annotation, expert judgment, or keyword matching—methods that are labor-intensive, error-prone, and fail to capture semantic richness [@Boselli2018]. LAiSER fills this gap by providing an accessible, flexible, and accurate framework for automated skill extraction and taxonomy alignment, enabling researchers, educators, and workforce professionals to analyze skill demands at scale.
 
-## 1. Introduction
+# Key Features
 
-### 1.1 Motivation
+LAiSER provides three core innovations:
 
-The contemporary labor market is characterized by rapid technological change, evolving skill requirements, and a growing disconnect between educational curricula and industry needs [@Autor2013]. This misalignment creates significant challenges for multiple stakeholders: learners struggle to identify relevant competencies, educators lack precise understanding of labor market demands, and employers face difficulties in communicating skill requirements effectively. A fundamental barrier to addressing these challenges is the absence of a standardized, computationally tractable representation of skills that can bridge the semantic gap between different domains and stakeholders.
+1. **Intelligent Text Preprocessing**: Domain-aware preprocessing removes extraneous information (company branding, legal boilerplate, benefits) while preserving task-relevant skill content.
 
[…]
