### 2025-12-08 e8ca787
Enhance skill alignment functionality by adding similarity threshold and top_k parameters to align_skills and align_skills_to_taxonomy methods. Update documentation to reflect new parameters and their defaults.
