### 2026-01-04 2429bed
Few shot prompt enhancement
