### 2026-01-04 3e2b069
Revert "Few shot prompt enhancement"
