### 2026-01-04 c69be39
Version bump and bug fixes
