### 2026-01-16 #9336 @crvernon
@atrisovic and @phanindra-max - Since this submission has not yet it will need to be adjusted for the new JOSS policy.  See the following:

# Update regarding JOSS scope and paper format changes

## What this means for your submission

For full details, please see:
* Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
* An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html

We understand this may be frustrating, particularly if you've been waiting for some time. We apologize for that. These changes reflect the rapidly evolving landscape of software development and our commitment to ensuring JOSS publications recognize genuine scholarly contributions.
