### 2026-02-04 e04bba9
feat(index): end-to-end FAISS index lifecycle
