+++ laiser/services.py
@@ -20,8 +20,7 @@ from laiser.config import (
 )
 from laiser.exceptions import SkillExtractionError, InvalidInputError
 from laiser.data_access import DataAccessLayer, FAISSIndexManager
-import faiss
-from pathlib import Path
+
 class PromptBuilder:
     """Builds prompts for different types of skill extraction tasks"""
     
@@ -203,7 +202,7 @@ class SkillAlignmentService:
     def __init__(self, data_access: DataAccessLayer, faiss_manager: FAISSIndexManager):
         self.data_access = data_access
         self.faiss_manager = faiss_manager
-    
+        self.faiss_manager.initialize_index(force_rebuild=False)
 
     
     def align_skills_to_taxonomy(
@@ -240,11 +239,7 @@ class SkillAlignmentService:
         skill_tags = []
         correlations = []
 
-        script_dir = Path(__file__).parent
-        local_index_path = script_dir / "public" / "skills_v03.index"
-        local_json_path = script_dir / "public" / "skills_df.json"
-        skill_df = pd.read_json(str(local_json_path))
-        index = faiss.read_index(str(local_index_path))
+
         model = self.data_access.get_embedding_model()
 
         # Get SkillLabel to SkillTag mapping
@@ -253,18 +248,29 @@ class SkillAlignmentService:
         for skill in raw_skills:
             query_vec = model.encode([skill], normalize_embeddings=True)
             # Search for the single best match for each raw skill
-            D, I = index.search(np.array(query_vec).astype('float32'), 1)
-            
+            results = self.faiss_manager.search_similar_skills(
+                np.array(query_vec).astype("float32"), top_k=1
+            )
+
+            if not results:
+                continue
+
+            best = results[0]
+            similarity = float(best.get("Similarity", 0.0))
+
             # Only include matches above the threshold
-            if D[0][0] >= similarity_threshold:
-                canonical_skill = skill_df.iloc[I[0][0]]["skill"]
+            if similarity >= similarity_threshold:
+                canonical_skill = str(best.get("Skill", "")).strip()
+                if not canonical_skill:
+                    continue
+
                 mapped_skills.append(canonical_skill)
                 raw_skills_matched.append(skill)
 
                 # Get the corresponding SkillTag
                 skill_tag = label_to_tag_mapping.get(canonical_skill, "")
                 skill_tags.append(skill_tag)
-                correlations.append(float(D[0][0]))
+                correlations.append(similarity)
 
         # Apply top_k limit: sort by correlation (descending) and take top_k
         if len(mapped_skills) > top_k:
@@ -298,15 +304,51 @@ class SkillExtractionService:
     """Main service for skill extraction operations"""
     
     def __init__(self):
+        print("Initalisng skill extraction service")
         self.data_access = DataAccessLayer()
         self.faiss_manager = FAISSIndexManager(self.data_access)
         self.alignment_service = SkillAlignmentService(self.data_access, self.faiss_manager)
         self.prompt_builder = PromptBuilder()
         self.response_parser = ResponseParser()
-        
         # Initialize FAISS index
-        self.faiss_manager.initialize_index()
+        self.faiss_manager.initialize_index(force_rebuild=False)
+        print("done Initalisng skill extraction service")
     
+    def align_extracted_skills(self, raw_skills: List[str], document_id: str = '0',description: str = '',similarity_threshold: float = 0.20,top_k: int = DEFAULT_TOP_K) -> pd.DataFrame:
+        """
+        Align extracted skills to taxonomy.
+        
+        Parameters
+        ----------
+        raw_skills : List[str]
+            List of raw extracted skills
+        document_id : str
+            Document identifier
+        description : str
+            Full description text for context
+        similarity_threshold : float
[…]
