+++ laiser/data_access.py
@@ -67,6 +67,7 @@ from sentence_transformers import SentenceTransformer
 
 from laiser.config import (
     ESCO_SKILLS_URL, 
+    OSN_SKILLS_URL,
     COMBINED_SKILLS_URL, 
     FAISS_INDEX_URL, 
     DEFAULT_EMBEDDING_MODEL
@@ -80,6 +81,7 @@ class DataAccessLayer:
     def __init__(self):
         self.embedding_model = None
         self._esco_df = None
+        self._osn_df = None
         self._combined_df = None
     
     def get_embedding_model(self) -> SentenceTransformer:
@@ -92,17 +94,24 @@ class DataAccessLayer:
         """Load ESCO skills taxonomy data"""
         if self._esco_df is None:
             try:
-                print("Loading ESCO skill taxonomy data...")
                 self._esco_df = pd.read_csv(ESCO_SKILLS_URL)
             except Exception as e:
                 raise LAiSERError(f"Failed to load ESCO skills data: {e}")
         return self._esco_df
     
+    def load_osn_skills(self) -> pd.DataFrame:
+        """Load ESCO skills taxonomy data"""
+        if self._osn_df is None:
+            try:
+                self._osn_df = pd.read_csv(OSN_SKILLS_URL)
+            except Exception as e:
+                raise LAiSERError(f"Failed to load OSN skills data: {e}")
+        return self._osn_df
+    
     def load_combined_skills(self) -> pd.DataFrame:
         """Load combined skills taxonomy data"""
         if self._combined_df is None:
             try:
-                print("Loading combined skill taxonomy data...")
                 self._combined_df = pd.read_csv(COMBINED_SKILLS_URL)
             except Exception as e:
                 raise LAiSERError(f"Failed to load combined skills data: {e}")
@@ -140,19 +149,15 @@ class DataAccessLayer:
 
         return mapping
     
-    def build_faiss_index(self, skill_names: List[str]) -> faiss.IndexFlatIP:
+    def build_faiss_index(self, text: List[str]) -> faiss.IndexFlatIP:
         """Build FAISS index for given skill names"""
         try:
-            print("Building FAISS index...")
             model = self.get_embedding_model()
-            embeddings = model.encode(skill_names, convert_to_numpy=True, show_progress_bar=True)
-            
+            embeddings = model.encode(text,normalize_embeddings=True, convert_to_numpy=True, show_progress_bar=True)
             dimension = embeddings.shape[1]
             index = faiss.IndexFlatIP(dimension)
-            faiss.normalize_L2(embeddings)
-            index.add(embeddings)
-            
-            return index
+            index.add(np.asarray(embeddings, dtype='float32'))
+            return index, embeddings 
         except Exception as e:
             raise FAISSIndexError(f"Failed to build FAISS index: {e}")
     
@@ -161,7 +166,6 @@ class DataAccessLayer:
         try:
             os.makedirs(os.path.dirname(file_path), exist_ok=True)
             faiss.write_index(index, file_path)
-            print(f"FAISS index saved to {file_path}")
         except Exception as e:
             raise FAISSIndexError(f"Failed to save FAISS index: {e}")
     
@@ -169,7 +173,6 @@ class DataAccessLayer:
         """Load FAISS index from file"""
         try:
             if os.path.exists(file_path):
-                print(f"Loading FAISS index from {file_path}...")
                 return faiss.read_index(file_path)
             return None
         except Exception as e:
@@ -178,7 +181,6 @@ class DataAccessLayer:
     def download_faiss_index(self, url: str, local_path: str) -> bool:
         """Download FAISS index from URL"""
         try:
-            print(f"Downloading FAISS index from {url}...")
[…]
