### 2026-02-04 dbab001
test(index): add FAISS index lifecycle test
