### 2026-02-05 269d282
feat(data-access): expose skill metadata via API
