+++ laiser/data_access.py
@@ -62,6 +62,7 @@ import pandas as pd
 import faiss
 import numpy as np
 from pathlib import Path
+import json
 from typing import Optional, List, Dict, Any
 from sentence_transformers import SentenceTransformer
 
@@ -74,6 +75,8 @@ from laiser.config import (
 )
 from laiser.exceptions import FAISSIndexError, LAiSERError
 
+import logging
+logger = logging.getLogger(__name__)
 
 class DataAccessLayer:
     """Handles data loading and external API calls"""
@@ -82,6 +85,7 @@ class DataAccessLayer:
         self.embedding_model = None
         self._esco_df = None
         self._osn_df = None
+
         self._combined_df = None
     
     def get_embedding_model(self) -> SentenceTransformer:
@@ -100,7 +104,7 @@ class DataAccessLayer:
         return self._esco_df
     
     def load_osn_skills(self) -> pd.DataFrame:
-        """Load ESCO skills taxonomy data"""
+        """Load OSN skills taxonomy data"""
         if self._osn_df is None:
             try:
                 self._osn_df = pd.read_csv(OSN_SKILLS_URL)
@@ -108,6 +112,26 @@ class DataAccessLayer:
                 raise LAiSERError(f"Failed to load OSN skills data: {e}")
         return self._osn_df
     
+    def load_skill_metadata(self,file_path: str) -> pd.DataFrame:
+        """Load skills metadata JSON generated during FAISS index build"""
+        if self._combined_df is None:
+            try:
+                if not os.path.exists(file_path):
+                    raise FileNotFoundError(
+                        f"Skills metadata file not found at {file_path}. "
+                        "Build or download the FAISS index first."
+                    )
+
+                self._combined_df = pd.read_json(
+                    file_path,
+                    orient="records"
+                )
+            except Exception as e:
+                raise LAiSERError(f"Failed to load skills metadata: {e}")
+
+        return self._combined_df
+
+
     def load_combined_skills(self) -> pd.DataFrame:
         """Load combined skills taxonomy data"""
         if self._combined_df is None:
@@ -169,6 +193,17 @@ class DataAccessLayer:
         except Exception as e:
             raise FAISSIndexError(f"Failed to save FAISS index: {e}")
     
+    def save_skill_metadata_json(self, metadata: list, file_path: str) -> None:
+        """Save skills metadata to JSON file"""
+        try:
+            os.makedirs(os.path.dirname(file_path), exist_ok=True)
+
+            with open(file_path, "w", encoding="utf-8") as f:
+                json.dump(metadata, f, indent=2, ensure_ascii=False)
+
+        except Exception as e:
+            raise LAiSERError(f"Failed to save skills metadata: {e}")
+
     def load_faiss_index(self, file_path: str) -> Optional[faiss.IndexFlatIP]:
         """Load FAISS index from file"""
         try:
@@ -206,64 +241,77 @@ class FAISSIndexManager:
         self.embeddings = None
         self.metadata = None
     
-    def initialize_index(self, force_rebuild: bool = False) -> faiss.IndexFlatIP:
+    # Issue: Do we even need this? Can't this be done in init
+    def initialize_index(self, force_rebuild: bool = False, debug: bool = True ) -> faiss.IndexFlatIP:
         """Initialize FAISS index (load or build)"""
-        # Define paths
+        # [GFI_HelloWorld]: config this not hardcode
         script_dir = Path(__file__).parent
         local_index_path = script_dir / "public" / "skills_v04.index"
         local_json_path = script_dir / "public" / "skills_df.json" 
         local_npy_path = script_dir / "public" / "skill_embeddings.npy"
-        # Try to load existing index
+ 
         if not force_rebuild:
-            self.index = self.data_access.load_faiss_index(str(local_index_path))
             
-            if self.index is None:
[…]
