### 2026-02-05 aaeccbb
feat(alignment): add taxonomy metadata to skill mapping
