+++ laiser/services.py
@@ -21,6 +21,9 @@ from laiser.config import (
 from laiser.exceptions import SkillExtractionError, InvalidInputError
 from laiser.data_access import DataAccessLayer, FAISSIndexManager
 
+import logging
+logger = logging.getLogger(__name__)
+
 class PromptBuilder:
     """Builds prompts for different types of skill extraction tasks"""
     
@@ -204,7 +207,6 @@ class SkillAlignmentService:
         self.faiss_manager = faiss_manager
         self.faiss_manager.initialize_index(force_rebuild=False)
 
-    
     def align_skills_to_taxonomy(
         self,
         raw_skills: List[str],
@@ -212,94 +214,113 @@ class SkillAlignmentService:
         description: str = '',
         similarity_threshold: float = 0.20,
         top_k: int = DEFAULT_TOP_K,
+        debug: bool = False
     ) -> pd.DataFrame:
-        """
-        Align raw skills to taxonomy using the FAISS index.
-        
-        Parameters
-        ----------
-        raw_skills : List[str]
-            List of raw extracted skills to align
-        document_id : str
-            Document identifier
-        description : str
-            Full description text for context
-        similarity_threshold : float
-            Minimum similarity score for a match to be included (default: 0.20)
-        top_k : int
-            Maximum number of aligned skills to return (default: 25)
-        
-        Returns
-        -------
-        pd.DataFrame
-            DataFrame with aligned skills, limited to top_k results above threshold
-        """
         mapped_skills = []
         raw_skills_matched = []
-        skill_tags = []
+        taxonomy_descriptions = []
+        taxonomy_sources = []
         correlations = []
 
+        def log_debug(msg: str):
+            if debug:
+                logger.debug(msg)
+
+        log_debug(f"[align] raw_skills={len(raw_skills)} threshold={similarity_threshold} top_k={top_k}")
 
         model = self.data_access.get_embedding_model()
 
-        # Get SkillLabel to SkillTag mapping
-        label_to_tag_mapping = self.data_access.get_skill_label_to_tag_mapping()
+        # metadata loaded once
+        metadata = self.faiss_manager.get_metadata()
+        log_debug(f"[align] metadata type={type(metadata).__name__} len={len(metadata)}")
+        if isinstance(metadata, pd.DataFrame) and not metadata.empty:
+            log_debug(f"[align] metadata columns={list(metadata.columns)}")
 
-        for skill in raw_skills:
+        for i, skill in enumerate(raw_skills):
+            log_debug(f"[skill {i}] raw='{skill}'")
+
             query_vec = model.encode([skill], normalize_embeddings=True)
-            # Search for the single best match for each raw skill
+
             results = self.faiss_manager.search_similar_skills(
                 np.array(query_vec).astype("float32"), top_k=1
             )
+            log_debug(f"[skill {i}] results={results}")
 
             if not results:
+                log_debug(f"[skill {i}] no results -> skip")
                 continue
 
             best = results[0]
             similarity = float(best.get("Similarity", 0.0))
+            meta_idx = best.get("Index")
+            canonical_skill = str(best.get("Skill", "")).strip()
 
-            # Only include matches above the threshold
-            if similarity >= similarity_threshold:
-                canonical_skill = str(best.get("Skill", "")).strip()
-                if not canonical_skill:
-                    continue
+            log_debug(f"[skill {i}] best='{canonical_skill}' sim={similarity:.4f} meta_idx={meta_idx}")
 
-                mapped_skills.append(canonical_skill)
-                raw_skills_matched.append(skill)
+            if similarity < similarity_threshold:
+                log_debug(f"[skill {i}] below threshold -> skip")
+                continue
 
-                # Get the corresponding SkillTag
[…]
