### 2026-02-05 af07dd3
chore(alignment_test): pytest files for alignment testing pipeline
