+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+markers =
+    alignment: minimal cost demo tests for alignment module
+    index: faiss index
