### 2026-02-05 8334dea
chore(combined.csv): combined csv for all skill taxonomy
