### 2026-02-05 ab48dc1
chore(skills_metadata): minor fixes
