### 2026-02-06 4a75124
test: core skill extractor
