+++ pytest.ini
@@ -2,3 +2,4 @@
 markers =
     alignment: minimal cost demo tests for alignment module
     index: faiss index
+    library: entire library
