### 2026-02-06 17dbfa5
test: remove legacy extract_and_align script
