+++ pytest.ini
@@ -3,3 +3,6 @@ markers =
     alignment: minimal cost demo tests for alignment module
     index: faiss index
     library: entire library
+    llm_cloud: cloud llms test
+    openai: openai
+    anthropic: anthropic
