+++ laiser/services.py
@@ -18,6 +18,8 @@ from laiser.config import (
     KSA_EXTRACTION_PROMPT,
     KSA_DETAILS_PROMPT
 )
+from laiser.llm_models.llm_router import LLMRouter
+from laiser.config import DEFAULT_BATCH_SIZE, DEFAULT_TOP_K
 from laiser.exceptions import SkillExtractionError, InvalidInputError
 from laiser.data_access import DataAccessLayer, FAISSIndexManager
 
@@ -28,10 +30,11 @@ class PromptBuilder:
     """Builds prompts for different types of skill extraction tasks"""
     
     @staticmethod
-    def build_skill_extraction_prompt(input_text: Dict[str, str], input_type: str) -> str:
+    def build_skill_extraction_prompt(input_text: str, input_type: str) -> str:
         """Build prompt for basic skill extraction"""
         if input_type == "job_desc":
-            return SKILL_EXTRACTION_PROMPT_JOB.format(query=input_text.get("description", ""))
+            extraction_prompt = SKILL_EXTRACTION_PROMPT_JOB.format(description=input_text)
+            return extraction_prompt
         elif input_type == "syllabus":
             return SKILL_EXTRACTION_PROMPT_SYLLABUS.format(
                 description=input_text.get("description", ""),
@@ -83,11 +86,98 @@ class PromptBuilder:
             num_key_kr=num_key_kr,
             num_key_tas=num_key_tas
         )
+    
+    def strong_preprocessing_prompt(self,raw_description):
+        prompt = f"""
+    You are a data preprocessing assistant trained to clean job descriptions for skill extraction.
 
+    Your task is to remove the following from the text:
+    - Company names, slogans, branding language
+    - Locations, phone numbers, email addresses, URLs
+    - Salary information, job ID, dates, scheduling info (e.g. 9am-5pm, weekends required)
+    - HR/legal boilerplate (EEO, diversity statements, veteran status, disability policies)
+    - Culture fluff like "fun environment", "fast-paced", "initiative", "self-motivated", "join us", "own your tomorrow", "apply now"
+    - Internal team names or product names (e.g. ACE, THD, IMT)
+    - Benefits sections (e.g. health & wellness, sabbatical, 401k, maternity, vacation)
+
+    Your output must *only retain the task-related job duties, technical responsibilities, required skills, qualifications, and tools* without rephrasing.
+
+    Input:
+    \"\"\"
+    {raw_description}
+    \"\"\"
+
+    Return only the cleaned job description.
+    ### CLEANED JOB DESCRIPTION:
+    """
+
+    ## ISSUE: Fix llm router params
+        response = llm_router(prompt, self.model_id, self.use_gpu, self.llm, 
+                                self.tokenizer, self.model, self.api_key)
+        cleaned = response.split("### CLEANED JOB DESCRIPTION:")[-1].strip()
+        return cleaned
+      
 
 class ResponseParser:
     """Parses responses from LLM models"""
     
+    @staticmethod
+    def _parse_skills_from_response(response: str) -> List[str]:
+        if not response or not response.strip():
+            return []
+
+        fragments: List[str] = []
+        stripped = response.strip()
+
+        code_match = re.search(r"```(?:json)?\s*(.*?)\s*```", stripped, re.DOTALL)
+        if code_match:
+            fragments.append(code_match.group(1).strip())
+
+        brace_match = re.search(r"\{.*?\}", stripped, re.DOTALL)
+        if brace_match:
+            fragments.append(brace_match.group(0).strip())
+
+        list_match = re.search(r"\[.*?\]", stripped, re.DOTALL)
+        if list_match:
+            fragments.append(list_match.group(0).strip())
+
+        fragments.append(stripped)
+
+        seen = set()
+        for fragment in fragments:
[…]
