+++ laiser/__init__.py
@@ -4,19 +4,18 @@ LAiSER - Leveraging Artificial Intelligence for Skills Extraction and Research
 A Python package for extracting and aligning skills from text using AI models.
 """
 
-__version__ = "0.3.2"
+__version__ = "0.3.3.6"
 
 # Import main classes for easy access
 try:
-    from .skill_extractor import Skill_Extractor
     from .skill_extractor_refactored import SkillExtractorRefactored
     
     # Make both available
-    __all__ = ['Skill_Extractor', 'SkillExtractorRefactored']
+    __all__ = [ 'SkillExtractorRefactored']
 except ImportError as e:
     # Handle cases where dependencies might not be available
     import warnings
-    warnings.warn(f"Could not import all LAiSER components: {e}")
+    warnings.warn(f"Could not import LAiSER components: {e}")
     __all__ = []
 
 # For backward compatibility
