+++ laiser/skill_extractor.py
@@ -1,665 +1,0 @@
-"""
-Module Description:
--------------------
-Class to extract skills from text and align them to existing taxonomy
-
-Ownership:
-----------
-Project: Leveraging Artificial intelligence for Skills Extraction and Research (LAiSER)
-Owner:  George Washington University Institute of Public Policy
-        Program on Skills, Credentials and Workforce Policy
-        Media and Public Affairs Building
-        805 21st Street NW
-        Washington, DC 20052
-        PSCWP@gwu.edu
-        https://gwipp.gwu.edu/program-skills-credentials-workforce-policy-pscwp
-
-License:
---------
-Copyright 2025 George Washington University Institute of Public Policy
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
-documentation files (the "Software"), to deal in the Software without restriction, including without limitation
-the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
-and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all copies or substantial portions of the
-Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
-WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
-OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-
-Input Requirements:
--------------------
-- Pandas Dataframe with ID and Text Column
-
-Output/Return Format:
-----------------------------
-- Pandas dataframe with below columns:
-    - "Research ID": text_id
-    - "Skill Name": Raw skill extracted,
-    - "Skill Tag": skill tag from taxonomy,
-    - "Correlation Coefficient": similarity_score
-
-
-"""
-"""
-Revision History:
------------------
-Rev No.     Date            Author              Description
-[1.0.0]     05/30/2024      Vedant M.           Initial Version
-[1.0.1]     06/01/2024      Vedant M.           Referencing utils.py and params.py
-[1.0.2]     06/08/2024      Satya Phanindra K.  Modify get_aligned_skills function to JSON output
-[1.0.3]     06/10/2024      Vedant M.           Updated functions extract_raw and align_skills for input and output
-[1.0.4]     06/13/2024      Vedant M.           Added function extractor to encapsulate both functions
-[1.0.5]     06/15/2024      Satya Phanindra K.  Replaced OpenAI API with HuggingFace API for skill extraction
-[1.0.6]     06/20/2024      Satya Phanindra K.  Added function to extract skills from text using Fine-Tuned Language Model's API
-[1.0.7]     07/03/2024      Satya Phanindra K.  Added CONDITIONAL GPU support for Fine-Tuned Language Model and error handling
-[1.0.9]     07/08/2024      Satya Phanindra K.  Added support for SkillNer model for skill extraction, if GPU not available
-[1.0.8]     07/11/2024      Satya Phanindra K.  Calculate cosine similarities in bulk for optimal performance.
[…]
