### 2026-02-20 132cb68
refactor(taxonomy mapping issues): index pipeline fix for mapping
