### 2026-02-25 34b9f5c
task: version update
