### 2026-02-26 ccd7d6c
fix(embedding_model): load embedding model during initalization
