### 2026-02-26 3aea697
feat: add LlamaCpp backend support and enhance skill extraction prompts. update: Changed gemini model being used from 2.5-pro to 2.5-flash
