+++ laiser/skill_extractor_refactored.py
@@ -145,7 +145,6 @@ class SkillExtractorRefactored:
         pd.DataFrame
             DataFrame with extracted and aligned skills
         """
-        print("data is passed", data)
         return self.skill_service.extract_and_align_core(
         data=data,
         id_column=id_column,
