### 2026-03-06 ec6a28d
Update Paper for new JOSS standards
