+++ paper.bib
@@ -55,3 +55,42 @@
   note = {Accessed: 2025-11-01}
 }
 
+@misc{SkillNER2022,
+  author = {Aito, Anas},
+  title = {SkillNER: A rule-based NLP module to extract skills from text},
+  year = {2022},
+  url = {https://github.com/AnasAito/SkillNER},
+  note = {Accessed: 2026-03-06}
+}
+
+@misc{Nesta2023,
+  author = {Kerle, India and Gallagher, Elizabeth and Sleeman, Cath and Vines, Jack},
+  title = {The Skills Extractor Library},
+  year = {2023},
+  url = {https://github.com/nestauk/ojd_daps_skills},
+  note = {Accessed: 2026-03-06}
+}
+
+@misc{EscoExtractor2024,
+  author = {Petrakis, Konstantinos},
+  title = {esco-skill-extractor: Extract ESCO skills and ISCO occupations from text},
+  year = {2024},
+  url = {https://github.com/KonstantinosPetrakis/esco-skill-extractor},
+  note = {Accessed: 2026-03-06}
+}
+
+@misc{DCC2025,
+  author = {{Digital Credentials Consortium}},
+  title = {Digital Credentials Consortium and collaborators launch Gen AI Badge Authoring Tool Project},
+  year = {2025},
+  url = {https://openlearning.mit.edu/news/digital-credentials-consortium-and-collaborators-launch-gen-ai-badge-authoring-tool-project},
+  note = {Accessed: 2026-03-06}
+}
+
+@misc{QSReimagine2025,
+  author = {{QS Quacquarelli Symonds}},
+  title = {QS Reimagine Education Awards 2025: Category Winners},
+  year = {2025},
+  url = {https://www.reimagine-education.com/},
+  note = {Accessed: 2026-03-06}
+}
