+++ paper.md
@@ -23,57 +23,38 @@ bibliography: paper.bib
 
 # Summary
 
-LAiSER is an artificial intelligence framework for extracting, standardizing, and aligning skill information from unstructured text with established skill taxonomies. The system addresses the lack of standardized, machine-readable skill information needed to facilitate communication between learners, educators, and employers. LAiSER employs a two-stage pipeline combining large language models (LLMs) with semantic vector search using FAISS [@Johnson2019] for high-precision skill extraction and taxonomy alignment to standards like ESCO [@ESCO2020]. The framework supports multiple computational backends—GPU-accelerated inference, cloud APIs, and CPU-only environments—making it accessible across diverse research and operational contexts.
+LAiSER is an open-source Python framework for extracting, standardizing, and aligning skill information from unstructured text with established skill taxonomies. The system addresses the lack of standardized, machine-readable skill information needed to facilitate communication between learners, educators, and employers. LAiSER employs a two-stage pipeline combining large language models (LLMs) with semantic vector search using FAISS [@Johnson2019] for high-precision skill extraction and taxonomy alignment to standards such as ESCO [@ESCO2020]. The framework supports multiple computational backends, including GPU-accelerated inference, cloud APIs, and CPU-only environments, making it accessible across diverse research and operational contexts. LAiSER also supports extraction within the Knowledge, Skills, and Abilities (KSA) framework with automatic proficiency level classification using the Scottish Credit and Qualifications Framework (SCQF) [@SCQF2019]. The source code is available at [https://github.com/LAiSER-Software/extract-module](https://github.com/LAiSER-Software/extract-module) under the MIT License, installable via `pip install laiser[gpu]` or `pip install laiser[cpu]`.
 
 # Statement of need
 
-The contemporary labor market is characterized by rapid technological change, evolving skill requirements, and growing disconnect between educational curricula and industry needs [@Autor2013]. Traditional skill extraction approaches rely on manual annotation, expert judgment, or keyword matching—methods that are labor-intensive, error-prone, and fail to capture semantic richness [@Boselli2018]. LAiSER fills this gap by providing an accessible, flexible, and accurate framework for automated skill extraction and taxonomy alignment, enabling researchers, educators, and workforce professionals to analyze skill demands at scale.
+The contemporary labor market is characterized by rapid technological change, evolving skill requirements, and growing disconnect between educational curricula and industry needs [@Autor2013]. Traditional skill extraction approaches rely on manual annotation, expert judgment, or keyword matching, all of which are labor-intensive, error-prone, and fail to capture semantic richness [@Boselli2018]. The challenge is compounded by the absence of integrated tooling that combines skill identification with alignment to recognized taxonomies, forcing researchers to assemble fragile multi-tool pipelines. LAiSER fills this gap by providing a unified, flexible, and accurate framework for automated skill extraction and taxonomy alignment, enabling researchers, educators, and workforce professionals to analyze skill demands at scale. Target users include labor economists studying occupational transitions, curriculum designers auditing program alignment with industry needs, and credentialing organizations producing standards-aligned digital badges.
 
-# Key Features
+# State of the field
 
-LAiSER provides three core innovations:
[…]
