### 2026-03-06 #9336 @phanindra-max
@crvernon and @atrisovic, I've updated the paper following the new format and a couple more updates. Thanks for your patience, looking forward for next steps :)
