### 2026-03-12 8f760e5
chore: Remove unused spaCy/SkillNer deps
