+++ pyproject.toml
@@ -34,11 +34,8 @@ dependencies = [
     "numpy>=1.21.0",
     "sentence-transformers>=2.2.0",
     "faiss-cpu>=1.7.0",
-    "spacy>=3.4.0",
+    "certifi>=2023.0.0",
     "requests>=2.25.0",
-    "tqdm>=4.62.0",
-    "skillNer==1.0.3",
-    "ipython",
     "psutil",
     "google-generativeai==0.8.5"
 ]
@@ -50,10 +47,12 @@ gpu = [
 dev = [
     "pytest>=6.0",
     "pytest-cov>=2.0",
+    "pytest-check>=2.0.0",
     "black>=22.0",
     "flake8>=4.0",
     "build>=0.8.0",
     "twine>=4.0.0",
+    "python-dotenv>=0.19.0,<1.0.0",
 ]
 
 [project.urls]
