### 2026-03-25 bc9ccba
chore: add GitHub Actions CI/CD pipeline
