+++ .github/workflows/pr-automation.yml
@@ -1,0 +1,119 @@
+name: PR Automation
+
+on:
+  pull_request:
+    types: [opened, reopened, synchronize]
+
+permissions:
+  pull-requests: write
+  contents: read
+
+jobs:
+  # ── 1. AI PR DESCRIPTION ──────────────────────────────────────────────────
+  ai-description:
+    name: AI PR Description
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Install dependencies
+        run: pip install google-generativeai requests
+
+      - name: Get PR diff
+        id: diff
+        run: |
+          git diff origin/${{ github.base_ref }}...HEAD > pr.diff
+          echo "diff<<EOF" >> $GITHUB_OUTPUT
+          cat pr.diff >> $GITHUB_OUTPUT
+          echo "EOF" >> $GITHUB_OUTPUT
+
+      - name: Generate and update PR description
+        env:
+          GEMINI_API_KEY: ${{ secrets.GEMINI_API_KEY }}
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          GITHUB_REPOSITORY: ${{ github.repository }}
+          PR_NUMBER: ${{ github.event.pull_request.number }}
+          PR_DIFF: ${{ steps.diff.outputs.diff }}
+        run: python .github/scripts/pr_description.py
+
+  # ── 2. REVIEW TIME LABEL ──────────────────────────────────────────────────
+  review-time-label:
+    name: Review Time Label
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - name: Calculate and label review time
+        uses: actions/github-script@v7
+        with:
+          script: |
+            const { data: pr } = await github.rest.pulls.get({
+              owner: context.repo.owner,
+              repo: context.repo.repo,
+              pull_number: context.issue.number,
+            });
+
+            const linesChanged = pr.additions + pr.deletions;
+
+            // ~200 lines per 5 minutes is a common benchmark
+            const minutes = Math.ceil(linesChanged / 40);
+
+            let label, color;
+            if (minutes < 5) {
+              label = `${minutes} min review`;
+              color = "0e8a16"; // green
+            } else if (minutes < 20) {
+              label = `${minutes} min review`;
+              color = "fbca04"; // yellow
+            } else {
+              label = `${minutes} min review`;
+              color = "b60205"; // red
+            }
+
+            // Create label if it doesn't exist
+            try {
+              await github.rest.issues.createLabel({
+                owner: context.repo.owner,
+                repo: context.repo.repo,
+                name: label,
+                color: color,
+              });
+            } catch (e) {
+              // label already exists, ignore
+            }
+
+            // Remove old review time labels
+            const { data: currentLabels } = await github.rest.issues.listLabelsOnIssue({
+              owner: context.repo.owner,
+              repo: context.repo.repo,
+              issue_number: context.issue.number,
+            });
+
+            for (const l of currentLabels) {
+              if (l.name.includes("min review")) {
+                await github.rest.issues.removeLabel({
+                  owner: context.repo.owner,
+                  repo: context.repo.repo,
+                  issue_number: context.issue.number,
+                  name: l.name,
+                });
+              }
+            }
+
+            // Add new label
+            await github.rest.issues.addLabels({
+              owner: context.repo.owner,
+              repo: context.repo.repo,
+              issue_number: context.issue.number,
+              labels: [label],
+            });
+
+            console.log(`Lines changed: ${linesChanged} → Label: "${label}"`);
