### 2026-03-25 865e891
style: apply black formatting
