+++ laiser/data_access.py
@@ -18,16 +18,16 @@ License:
 --------
 Copyright 2025 George Washington University Insitute of Public Policy
 
-Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files 
-(the “Software”), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, 
-merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is 
+Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files
+(the “Software”), to deal in the Software without restriction, including without limitation the rights to use, copy, modify,
+merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
 furnished to do so, subject to the following conditions:
 
 The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
 
-THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES 
-OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE 
-LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR 
+THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
 IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 
@@ -56,46 +56,47 @@ TODO:
 - 2: Implement caching for remote data sources.
 """
 
-import os
-import requests
 import io
+import json
+import logging
+import os
+from pathlib import Path
+from typing import Any, Dict, List, Optional
+
 import certifi
-import pandas as pd
 import faiss
 import numpy as np
-from pathlib import Path
-import json
-from typing import Optional, List, Dict, Any
+import pandas as pd
+import requests
 from sentence_transformers import SentenceTransformer
 
 from laiser.config import (
-    ESCO_SKILLS_URL, 
+    COMBINED_SKILLS_URL,
+    DEFAULT_EMBEDDING_MODEL,
+    ESCO_SKILLS_URL,
+    FAISS_INDEX_URL,
     OSN_SKILLS_URL,
-    COMBINED_SKILLS_URL, 
-    FAISS_INDEX_URL, 
-    DEFAULT_EMBEDDING_MODEL
 )
 from laiser.exceptions import FAISSIndexError, LAiSERError
 
-import logging
 logger = logging.getLogger(__name__)
+
 
 class DataAccessLayer:
     """Handles data loading and external API calls"""
-    
+
     def __init__(self):
         self.embedding_model = None
         self._esco_df = None
         self._osn_df = None
 
         self._combined_df = None
-    
+
     def get_embedding_model(self) -> SentenceTransformer:
         """Get or initialize the embedding model"""
[…]
