+++ laiser/services.py
@@ -4,71 +4,85 @@ Service layer for skill extraction and processing
 This module contains the core business logic for skill extraction.
 """
 
-import re
 import json
+import logging
+import re
+from typing import Any, Dict, List, Optional, Tuple, Union
+
 import numpy as np
-import torch
-from typing import List, Dict, Any, Tuple, Optional, Union
 import pandas as pd
+import torch
 
 from laiser.config import (
+    COMBINED_EXTRACTION_PROMPT,
+    DEFAULT_BATCH_SIZE,
     DEFAULT_TOP_K,
+    KSA_DETAILS_PROMPT,
+    KSA_EXTRACTION_PROMPT,
     SCQF_LEVELS,
     SKILL_EXTRACTION_PROMPT_JOB,
     SKILL_EXTRACTION_PROMPT_SYLLABUS,
-    KSA_EXTRACTION_PROMPT,
-    KSA_DETAILS_PROMPT,
-    COMBINED_EXTRACTION_PROMPT
 )
-from laiser.llm_models.llm_router import LLMRouter
-from laiser.config import DEFAULT_BATCH_SIZE, DEFAULT_TOP_K
-from laiser.exceptions import SkillExtractionError, InvalidInputError, LAiSERError
 from laiser.data_access import DataAccessLayer, FAISSIndexManager
+from laiser.exceptions import InvalidInputError, LAiSERError, SkillExtractionError
+from laiser.llm_models.llm_router import LLMRouter
 
-import logging
 logger = logging.getLogger(__name__)
+
 
 class PromptBuilder:
     """Builds prompts for different types of skill extraction tasks"""
-    
+
     @staticmethod
     def build_skill_extraction_prompt(input_text: str, input_type: str) -> str:
         """Build prompt for basic skill extraction"""
         if input_type == "job_desc":
-            extraction_prompt = SKILL_EXTRACTION_PROMPT_JOB.format(description=input_text)
+            extraction_prompt = SKILL_EXTRACTION_PROMPT_JOB.format(
+                description=input_text
+            )
             return extraction_prompt
         elif input_type == "syllabus":
             return SKILL_EXTRACTION_PROMPT_SYLLABUS.format(
                 description=input_text.get("description", ""),
-                learning_outcomes=input_text.get("learning_outcomes", "")
+                learning_outcomes=input_text.get("learning_outcomes", ""),
             )
         else:
             raise InvalidInputError(f"Unsupported input type: {input_type}")
-    
+
     @staticmethod
     def build_ksa_extraction_prompt(
-        query: Dict[str, str], 
-        input_type: str, 
+        query: Dict[str, str],
+        input_type: str,
         num_key_skills: int,
-        num_key_kr: str, 
+        num_key_kr: str,
         num_key_tas: str,
-        esco_skills: List[str] = None
+        esco_skills: List[str] = None,
     ) -> str:
         """Build prompt for KSA (Knowledge, Skills, Abilities) extraction"""
-        
-        input_desc = "job description" if input_type == "job_desc" else "course syllabus description and its learning outcomes"
-        
+
+        input_desc = (
+            "job description"
+            if input_type == "job_desc"
+            else "course syllabus description and its learning outcomes"
+        )
+
         if input_type == "syllabus":
             input_text = f"### Input:\\n**Course Description:** {query.get('description', '')}\\n**Learning Outcomes:** {query.get('learning_outcomes', '')}"
         else:
             input_text = f"### Input:\\n{query.get('description', '')}"
-        
+
         # Format SCQF levels
-        scqf_levels_text = "\\n".join([f"  - {level}: {desc}" for level, desc in SCQF_LEVELS.items()])
-        
+        scqf_levels_text = "\\n".join(
[…]

[…]
