### 2026-03-25 3fa46c1
ci: configure flake8 ignore rules
