### 2026-03-25 2a16f97
fix: apply flake8 compliance across codebase
