+++ tests/test_extract_and_align.py
@@ -1,6 +1,5 @@
 import os
 
-import pandas as pd
 import pytest
 from dotenv import load_dotenv
 
