+++ laiser/llm_models/__init__.py
@@ -23,5 +23,6 @@ __all__ = [
     "load_model_from_transformer",
     "LLMRouter",
     "GeminiAPI",
-    "HuggingFaceLLM" "LlamaCppBackend",
+    "HuggingFaceLLM",
+    "LlamaCppBackend",
 ]
