+++ laiser/services.py
@@ -36,9 +36,7 @@ class PromptBuilder:
     def build_skill_extraction_prompt(input_text: str, input_type: str) -> str:
         """Build prompt for basic skill extraction"""
         if input_type == "job_desc":
-            extraction_prompt = SKILL_EXTRACTION_PROMPT_JOB.format(
-                description=input_text
-            )
+            extraction_prompt = SKILL_EXTRACTION_PROMPT_JOB.format(description=input_text)
             return extraction_prompt
         elif input_type == "syllabus":
             return SKILL_EXTRACTION_PROMPT_SYLLABUS.format(
@@ -60,9 +58,7 @@ class PromptBuilder:
         """Build prompt for KSA (Knowledge, Skills, Abilities) extraction"""
 
         input_desc = (
-            "job description"
-            if input_type == "job_desc"
-            else "course syllabus description and its learning outcomes"
+            "job description" if input_type == "job_desc" else "course syllabus description and its learning outcomes"
         )
 
         if input_type == "syllabus":
@@ -71,16 +67,10 @@ class PromptBuilder:
             input_text = f"### Input:\\n{query.get('description', '')}"
 
         # Format SCQF levels
-        scqf_levels_text = "\\n".join(
-            [f"  - {level}: {desc}" for level, desc in SCQF_LEVELS.items()]
-        )
+        scqf_levels_text = "\\n".join([f"  - {level}: {desc}" for level, desc in SCQF_LEVELS.items()])
 
         # Prepare ESCO context
-        esco_context_block = (
-            ", ".join(esco_skills)
-            if esco_skills
-            else "No relevant skills found in taxonomy"
-        )
+        esco_context_block = ", ".join(esco_skills) if esco_skills else "No relevant skills found in taxonomy"
 
         return KSA_EXTRACTION_PROMPT.format(
             input_desc=input_desc,
@@ -93,9 +83,7 @@ class PromptBuilder:
         )
 
     @staticmethod
-    def build_ksa_details_prompt(
-        skill: str, description: str, num_key_kr: int = 3, num_key_tas: int = 3
-    ) -> str:
+    def build_ksa_details_prompt(skill: str, description: str, num_key_kr: int = 3, num_key_tas: int = 3) -> str:
         """Build prompt for getting detailed KSA information for a specific skill"""
         return KSA_DETAILS_PROMPT.format(
             skill=skill,
@@ -191,9 +179,7 @@ class ResponseParser:
             # Remove any unwanted prefixes and tags
             clean_lines = []
             for line in lines:
-                if line.startswith("<start_of_turn>") or line.startswith(
-                    "<end_of_turn>"
-                ):
+                if line.startswith("<start_of_turn>") or line.startswith("<end_of_turn>"):
                     continue
                 if "--" in line:  # Skip separator lines
                     continue
@@ -237,19 +223,13 @@ class ResponseParser:
                     )
                     if knowledge_match:
                         knowledge_raw = knowledge_match.group(1).strip()
-                        skill_data["Knowledge Required"] = [
-                            k.strip() for k in knowledge_raw.split(",") if k.strip()
-                        ]
+                        skill_data["Knowledge Required"] = [k.strip() for k in knowledge_raw.split(",") if k.strip()]
 
                     # Extract task abilities (multi-line support)
-                    task_match = re.search(
-                        r"Task Abilities:\s*(.*?)(?=\s*$)", item, re.DOTALL
-                    )
+                    task_match = re.search(r"Task Abilities:\s*(.*?)(?=\s*$)", item, re.DOTALL)
                     if task_match:
                         task_raw = task_match.group(1).strip()
-                        skill_data["Task Abilities"] = [
[…]
