+++ laiser/llm_models/llm_router.py
@@ -68,9 +68,7 @@ except ImportError as e:
     print(f"Warning: Gemini support not available: {e}")
 
     def gemini_generate(*args, **kwargs):
-        raise ImportError(
-            "Gemini support is not available. Please install google-generativeai package."
-        )
+        raise ImportError("Gemini support is not available. Please install google-generativeai package.")
 
 
 # Import with error handling for optional dependencies
@@ -85,16 +83,12 @@ except ImportError as e:
     print(f"Warning: HuggingFace LLM support not available: {e}")
 
     def llm_generate_vllm(*args, **kwargs):
-        raise ImportError(
-            "HuggingFace LLM support is not available. Please install required packages."
-        )
+        raise ImportError("HuggingFace LLM support is not available. Please install required packages.")
 
 
 class LLMRouter:
 
-    def __init__(
-        self, model_id: str, use_gpu: bool, hf_token=None, api_key=None, backend=None
-    ):
+    def __init__(self, model_id: str, use_gpu: bool, hf_token=None, api_key=None, backend=None):
         self.model_id = model_id
         self.use_gpu = use_gpu
         self.hf_token = hf_token
@@ -189,6 +183,4 @@ class LLMRouter:
         self.llm = load_model_from_vllm(self.model_id, self.hf_token)
 
     def _initialize_transformer(self):
-        self.tokenizer, self.model = load_model_from_transformer(
-            self.model_id, self.hf_token
-        )
+        self.tokenizer, self.model = load_model_from_transformer(self.model_id, self.hf_token)
