### 2026-03-25 5626c48
fix: pass PR diff as file to avoid argument list too long error
