### 2026-03-25 0885ef0
fix: pin HuggingFace model downloads to main revision
