+++ laiser/llm_models/model_loader.py
@@ -69,20 +69,22 @@ def load_model_from_transformer(model_id: str = None, token: str = ""):
     quantization_config = BitsAndBytesConfig(load_in_8bit=True)
 
     try:
-        tokenizer = AutoTokenizer.from_pretrained(model_id, use_auth_token=token)
+        tokenizer = AutoTokenizer.from_pretrained(model_id, use_auth_token=token, revision="main")
         model = AutoModelForCausalLM.from_pretrained(
             model_id,
             use_auth_token=token,
+            revision="main",
             quantization_config=quantization_config,
             device_map="auto",
         )
     except (RepositoryNotFoundError, EntryNotFoundError, OSError) as e:
         print(f"[WARN] Failed to load model '{model_id}': {e}")
         print(f"[INFO] Falling back to default model: {DEFAULT_TRANSFORMER_MODEL_ID}")
-        tokenizer = AutoTokenizer.from_pretrained(DEFAULT_TRANSFORMER_MODEL_ID, use_auth_token=token)
+        tokenizer = AutoTokenizer.from_pretrained(DEFAULT_TRANSFORMER_MODEL_ID, use_auth_token=token, revision="main")
         model = AutoModelForCausalLM.from_pretrained(
             DEFAULT_TRANSFORMER_MODEL_ID,
             use_auth_token=token,
+            revision="main",
             quantization_config=quantization_config,
             device_map="auto",
         )
