### 2026-03-25 707d6d9
ci: ignore CVE-2026-4539 in pip-audit until pygments releases a fix
