### 2026-03-25 d332834
fix: suppress bandit B615 warnings for pinned HuggingFace downloads
