### 2026-03-25 a187de8
feat: add CodeQL security scanning and Dependabot auto-updates
