### 2026-03-25 7e11588
ci: add type checking, license, dead code and complexity checks
