### 2026-03-25 06bf100
chore: add PR preview package deployment and additional CI checks
