### 2026-03-25 27c50d0
ci: add pr comments for complexity and hygiene, remove redundant deploy-testpypi
