+++ .github/workflows/ci.yml
@@ -97,6 +97,8 @@ jobs:
   dead-code:
     name: Dead Code
     runs-on: ubuntu-latest
+    permissions:
+      pull-requests: write
     steps:
       - uses: actions/checkout@v4
 
@@ -108,13 +110,69 @@ jobs:
         run: pip install vulture
 
       - name: vulture
-        run: vulture laiser/ --min-confidence 80
-        continue-on-error: true
+        id: vulture
+        run: |
+          set +e
+          OUTPUT=$(vulture laiser/ --min-confidence 80 2>&1)
+          EXIT_CODE=$?
+          echo "output<<EOF" >> $GITHUB_OUTPUT
+          echo "$OUTPUT" >> $GITHUB_OUTPUT
+          echo "EOF" >> $GITHUB_OUTPUT
+          echo "found=$( [ $EXIT_CODE -ne 0 ] && echo true || echo false )" >> $GITHUB_OUTPUT
+          exit 0
+
+      - name: Post code hygiene summary
+        uses: actions/github-script@v7
+        with:
+          script: |
+            const found = '${{ steps.vulture.outputs.found }}' === 'true';
+            const output = `${{ steps.vulture.outputs.output }}`.trim();
+
+            const lines = found ? output.split('\n') : [];
+            const items = lines.map(l => {
+              const m = l.match(/^(.+):(\d+):\s+(.+)$/);
+              return m ? `- \`${m[1]}:${m[2]}\` — ${m[3]}` : `- ${l}`;
+            }).slice(0, 25).join('\n');
+            const overflow = lines.length > 25 ? `\n_+${lines.length - 25} more_` : '';
+
+            const body = found
+              ? [
+                  '**Code Hygiene**',
+                  '',
+                  'The following items are defined but not yet in active use. Routine cleanup — no impact on functionality.',
+                  '',
+                  items + overflow,
+                ].join('\n')
+              : '**Code Hygiene** — All clear. Nothing to clean up.';
+
+            const { data: comments } = await github.rest.issues.listComments({
+              owner: context.repo.owner,
+              repo: context.repo.repo,
+              issue_number: context.issue.number,
+            });
+            const existing = comments.find(c => c.body.includes('**Code Hygiene**'));
+            if (existing) {
+              await github.rest.issues.updateComment({
+                owner: context.repo.owner,
+                repo: context.repo.repo,
+                comment_id: existing.id,
+                body,
+              });
+            } else {
+              await github.rest.issues.createComment({
+                owner: context.repo.owner,
+                repo: context.repo.repo,
+                issue_number: context.issue.number,
+                body,
+              });
+            }
 
   # ── 6. COMPLEXITY ──────────────────────────────────────────────────────────
   complexity:
     name: Complexity
     runs-on: ubuntu-latest
+    permissions:
+      pull-requests: write
     steps:
       - uses: actions/checkout@v4
 
@@ -126,8 +184,71 @@ jobs:
         run: pip install radon
 
       - name: radon complexity
-        run: radon cc laiser/ --min B --show-complexity --average
-        continue-on-error: true
+        id: radon
+        run: |
+          set +e
+          OUTPUT=$(radon cc laiser/ --min B --show-complexity --average 2>&1)
+          echo "output<<EOF" >> $GITHUB_OUTPUT
+          echo "$OUTPUT" >> $GITHUB_OUTPUT
+          echo "EOF" >> $GITHUB_OUTPUT
+          exit 0
+
+      - name: Post complexity insights
+        uses: actions/github-script@v7
+        with:
+          script: |
+            const output = `${{ steps.radon.outputs.output }}`.trim();
+            if (!output) return;
+
+            const lines = output.split('\n');
+            const avgLine = lines[lines.length - 1] || '';
[…]
