### 2026-03-25 d4eaace
ci: rename preview package to laiser-preview, sort complexity table by score
