+++ laiser/services.py
@@ -14,12 +14,13 @@ import pandas as pd
 import torch
 
 from laiser.config import (
+    COMBINED_EXTRACTION_PROMPT,
     DEFAULT_BATCH_SIZE,
     DEFAULT_TOP_K,
     KSA_DETAILS_PROMPT,
     KSA_EXTRACTION_PROMPT,
+    KT_FROM_SKILLS_PROMPT,
     SCQF_LEVELS,
-    SKILL_EXTRACTION_PROMPT_JOB,
     SKILL_EXTRACTION_PROMPT_SYLLABUS,
 )
 from laiser.data_access import DataAccessLayer, FAISSIndexManager
@@ -36,7 +37,7 @@ class PromptBuilder:
     def build_skill_extraction_prompt(input_text: str, input_type: str) -> str:
         """Build prompt for basic skill extraction"""
         if input_type == "job_desc":
-            extraction_prompt = SKILL_EXTRACTION_PROMPT_JOB.format(description=input_text)
+            extraction_prompt = COMBINED_EXTRACTION_PROMPT.format(description=input_text)
             return extraction_prompt
         elif input_type == "syllabus":
             return SKILL_EXTRACTION_PROMPT_SYLLABUS.format(
@@ -90,6 +91,20 @@ class PromptBuilder:
             description=description,
             num_key_kr=num_key_kr,
             num_key_tas=num_key_tas,
+        )
+
+    @staticmethod
+    def build_knowledge_task_prompt(description: str, extracted_skills: List[str]) -> str:
+        """
+        Build prompt for extracting Knowledge and Tasks (Call 2 of v0.5 pipeline).
+
+        Uses the full job description as context alongside already-extracted skills
+        so that Knowledge and Task outputs are specific to this role, not generic.
+        """
+        skills_formatted = "\n".join(f"- {s}" for s in extracted_skills)
+        return KT_FROM_SKILLS_PROMPT.format(
+            description=description,
+            skills=skills_formatted,
         )
 
     def strong_preprocessing_prompt(self, raw_description):
@@ -245,6 +260,49 @@ class ResponseParser:
             return []
 
     @staticmethod
+    @staticmethod
+    def parse_knowledge_task_response(response: str) -> List[Dict[str, Any]]:
+        """
+        Parse the JSON response from KT_FROM_SKILLS_PROMPT.
+
+        Returns a list of dicts, each with keys:
+            skill     str
+            knowledge List[str]
+            tasks     List[str]
+        """
+        try:
+            # Strip markdown code fences if present
+            cleaned = re.sub(r"```(?:json)?", "", response).strip().rstrip("`").strip()
+            parsed = json.loads(cleaned)
+            results = parsed.get("results", [])
+
+            validated = []
+            for item in results:
+                if not isinstance(item, dict):
+                    continue
+                skill = str(item.get("skill", "")).strip()
+                knowledge = item.get("knowledge", [])
+                tasks = item.get("tasks", [])
+
+                if not isinstance(knowledge, list):
+                    knowledge = [str(knowledge)] if knowledge else []
+                if not isinstance(tasks, list):
+                    tasks = [str(tasks)] if tasks else []
+
+                if skill:
+                    validated.append(
+                        {
+                            "skill": skill,
+                            "knowledge": [k.strip() for k in knowledge if k.strip()],
+                            "tasks": [t.strip() for t in tasks if t.strip()],
+                        }
+                    )
+            return validated
+
+        except Exception as e:
+            logger.warning(f"Failed to parse knowledge/task response: {e}")
+            return []
+
     def parse_ksa_details_response(response: str) -> Tuple[List[str], List[str]]:
         """Parse KSA details response"""
[…]
