### 2026-03-31 fc28643
feat(data_access): add base and concrete FAISS index managers for Knowledge and Task
Introduces _BaseTaxonomyFAISSIndexManager with shared logic for
loading cached indexes, building from CSV, and filtered similarity
search. KnowledgeFAISSIndexManager and TaskFAISSIndexManager extend
it, pointing to the v0.5 taxonomy files in laiser/public/.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
