+++ laiser/data_access.py
@@ -70,12 +70,7 @@ import pandas as pd
 import requests
 from sentence_transformers import SentenceTransformer
 
-from laiser.config import (
-    COMBINED_SKILLS_URL,
-    DEFAULT_EMBEDDING_MODEL,
-    ESCO_SKILLS_URL,
-    OSN_SKILLS_URL,
-)
+from laiser.config import COMBINED_SKILLS_URL, DEFAULT_EMBEDDING_MODEL, ESCO_SKILLS_URL, OSN_SKILLS_URL
 from laiser.exceptions import FAISSIndexError, LAiSERError
 
 logger = logging.getLogger(__name__)
@@ -675,3 +670,164 @@ class FAISSIndexManager:
 
         except Exception as e:
             raise FAISSIndexError(f"Failed to search similar skills: {repr(e)}")
+
+
+class _BaseTaxonomyFAISSIndexManager:
+    """
+    Shared base for Knowledge and Task FAISS index managers.
+
+    Subclasses must define:
+        INDEX_FILENAME   str  — e.g. "knowledge_v05.index"
+        META_FILENAME    str  — e.g. "knowledge_df.json"
+        CSV_FILENAME     str  — e.g. "knowledge_taxonomy.csv"
+        LABEL            str  — human-readable label for log messages ("Knowledge" / "Task")
+
+    The CSV produced by the pipeline scripts must contain at minimum:
+        name         — concept or task name
+        description  — definition / occupation context
+        taxonomy     — source tag (e.g. "onet_knowledge", "esco_task")
+    Additional columns (field, action_verb, etc.) are preserved in metadata.
+    """
+
+    INDEX_FILENAME: str = ""
+    META_FILENAME: str = ""
+    CSV_FILENAME: str = ""
+    LABEL: str = ""
+
+    def __init__(self, data_access: DataAccessLayer):
+        self.data_access = data_access
+        self.index: Optional[faiss.IndexFlatIP] = None
+        self.metadata: Optional[pd.DataFrame] = None
+        self._item_names: Optional[List[str]] = None
+
+    def initialize_index(self, force_rebuild: bool = False) -> None:
+        """Load existing index + metadata from public/, or build from CSV if missing."""
+        script_dir = Path(__file__).parent
+        index_path = script_dir / "public" / self.INDEX_FILENAME
+        meta_path = script_dir / "public" / self.META_FILENAME
+        csv_path = script_dir / "public" / self.CSV_FILENAME
+
+        # Try loading cached index
+        if not force_rebuild and index_path.exists() and meta_path.exists():
+            try:
+                self.index = self.data_access.load_faiss_index(str(index_path))
+                self.metadata = pd.read_json(str(meta_path), orient="records")
+                logger.info(f"[{self.LABEL}FAISSIndex] Loaded cached index ({len(self.metadata)} entries).")
+                return
+            except Exception as e:
+                logger.warning(f"[{self.LABEL}FAISSIndex] Cache load failed: {e}. Rebuilding from CSV.")
+
+        # Build from CSV
+        if not csv_path.exists():
+            logger.warning(
+                f"[{self.LABEL}FAISSIndex] Taxonomy CSV not found at {csv_path}. "
+                f"Run scripts/build_{self.LABEL.lower()}_index.py to generate it. "
+                f"Alignment for {self.LABEL} will be unavailable."
+            )
+            return
+
+        df = pd.read_csv(str(csv_path), dtype=str)
+        df = df.loc[:, ~df.columns.str.match(r"^Unnamed")]
+
+        # Normalize required columns
+        for col in ("name", "description", "taxonomy"):
+            if col not in df.columns:
+                raise LAiSERError(
[…]
