### 2026-03-31 d3735f5
feat(config): add per-type similarity thresholds and improve KST prompt rules
Adds DEFAULT_SIMILARITY_THRESHOLDS dict with differentiated thresholds
per type (skill=0.20, knowledge=0.45, task=0.55). Rewrites the KST
prompt STEP 2 section with strict structured rules, explicit GOOD/BAD
examples, banned verb patterns, and word-count requirements to produce
higher-quality Knowledge and Task extractions.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
