### 2026-03-31 3c7e3ca
feat(services): generalize AlignmentService and wire KST indexes with ENABLES edges
Renames SkillAlignmentService to AlignmentService (generic, works with
any FAISS manager; old name kept as alias). Wires KnowledgeFAISSIndexManager
and TaskFAISSIndexManager into SkillExtractionService. Replaces TODO stubs
in knowledge/task alignment methods with real FAISS lookups and fallbacks.
Adds return_edges param and _derive_enables_edges() to derive Knowledge
ENABLES Task edges from LLM co-occurrence per skill.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
