+++ laiser/services.py
@@ -16,6 +16,7 @@ import torch
 from laiser.config import (
     COMBINED_EXTRACTION_PROMPT,
     DEFAULT_BATCH_SIZE,
+    DEFAULT_SIMILARITY_THRESHOLDS,
     DEFAULT_TOP_K,
     KSA_DETAILS_PROMPT,
     KSA_EXTRACTION_PROMPT,
@@ -23,7 +24,7 @@ from laiser.config import (
     SCQF_LEVELS,
     SKILL_EXTRACTION_PROMPT_SYLLABUS,
 )
-from laiser.data_access import DataAccessLayer, FAISSIndexManager
+from laiser.data_access import DataAccessLayer, FAISSIndexManager, KnowledgeFAISSIndexManager, TaskFAISSIndexManager
 from laiser.exceptions import InvalidInputError, LAiSERError
 from laiser.llm_models.llm_router import LLMRouter
 
@@ -329,14 +330,151 @@ class ResponseParser:
             return [], []
 
 
-class SkillAlignmentService:
-    """Service for aligning extracted skills with taxonomies"""
+class AlignmentService:
+    """
+    Generalized alignment service for Skills, Knowledge, and Tasks.
 
-    def __init__(self, data_access: DataAccessLayer, faiss_manager: FAISSIndexManager):
+    Uses any FAISS index manager that exposes:
+        .get_metadata() -> pd.DataFrame
+        .search_similar(query_vec, top_k, allowed_sources) -> List[Dict]  (key "Name")
+        .search_similar_skills(...)  (legacy key "Skill", for FAISSIndexManager)
+    """
+
+    def __init__(self, data_access: DataAccessLayer, faiss_manager):
         self.data_access = data_access
         self.faiss_manager = faiss_manager
-        self.faiss_manager.initialize_index(force_rebuild=False)
+        # FAISSIndexManager.initialize_index is a no-op if already loaded
+        if hasattr(self.faiss_manager, "initialize_index"):
+            self.faiss_manager.initialize_index(force_rebuild=False)
 
+    def align(
+        self,
+        raw_items: List[str],
+        document_id: str = "0",
+        description: str = "",
+        similarity_threshold: float = 0.20,
+        top_k: int = DEFAULT_TOP_K,
+        raw_col: str = "Raw Item",
+        taxonomy_col: str = "Taxonomy Item",
+        allowed_sources: Optional[List[str]] = None,
+        debug: bool = False,
+    ) -> pd.DataFrame:
+        """
+        Align a list of raw extracted strings to the loaded FAISS taxonomy.
+
+        Parameters
+        ----------
+        raw_col : str
+            Name for the "raw extracted" column in the output DataFrame.
+        taxonomy_col : str
+            Name for the "matched taxonomy" column in the output DataFrame.
+
+        Returns
+        -------
+        pd.DataFrame with columns: Research ID, raw_col, taxonomy_col,
+            Taxonomy Description, Taxonomy Source, Correlation Coefficient
+        """
+        mapped_items: List[str] = []
+        raw_matched: List[str] = []
+        taxonomy_descriptions: List[str] = []
+        taxonomy_sources: List[str] = []
+        correlations: List[float] = []
+
+        def log_debug(msg: str):
+            if debug:
+                logger.debug(msg)
+
+        log_debug(f"[align] raw_items={len(raw_items)} threshold={similarity_threshold} top_k={top_k}")
+
+        model = self.data_access.get_embedding_model()
+        try:
+            metadata = self.faiss_manager.get_metadata()
+        except Exception:
+            # Index not yet available — return empty DataFrame
+            return pd.DataFrame(
+                {
+                    "Research ID": pd.Series([], dtype=str),
+                    raw_col: [],
[…]
