### 2026-03-31 20303b7
test: add ENABLES edge derivation test for KST extraction
Adds test_kst_extraction_with_enables_edges to verify that return_edges=True
returns a dict with nodes and edges DataFrames, validates required edge
columns, and asserts all edges have Edge Type == ENABLES.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
