### 2026-03-31 9cebbcf
feat(scripts): add pipeline build scripts for Knowledge and Task indexes
Adds four scripts:
- build_knowledge_index.py: fetches O*NET Knowledge + ESCO knowledge
  entries, deduplicates, and builds the knowledge_v05.index FAISS index
- build_task_index.py: fetches O*NET Task Statements + ESCO task entries,
  deduplicates, and builds the tasks_v05.index FAISS index
- diagnose_raw_kt.py: inspects raw LLM KT extraction quality
- diagnose_taxonomy_coverage.py: checks taxonomy coverage and alignment gaps

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
