+++ scripts/build_knowledge_index.py
@@ -1,0 +1,248 @@
+"""
+Build the Knowledge FAISS index for LAiSER v0.5.
+
+Sources:
+  1. O*NET Knowledge (Knowledge.txt + Content Model Reference.txt)
+     Downloaded from: https://www.onetcenter.org/database.html#individual-files
+  2. ESCO Knowledge entries filtered from the ESCO skills CSV
+
+Output:
+  laiser/public/knowledge_taxonomy.csv  — unified knowledge taxonomy
+  laiser/public/knowledge_v05.index     — FAISS IndexFlatIP
+  laiser/public/knowledge_df.json       — metadata JSON
+
+Usage:
+  python scripts/build_knowledge_index.py
+  python scripts/build_knowledge_index.py --force-rebuild
+  python scripts/build_knowledge_index.py --onet-dir /path/to/onet_db_28_0/
+"""
+
+import argparse
+import io
+import logging
+import sys
+from pathlib import Path
+
+import certifi
+import pandas as pd
+import requests
+
+# Ensure the repo root is on sys.path so `laiser` can be imported
+REPO_ROOT = Path(__file__).resolve().parent.parent
+sys.path.insert(0, str(REPO_ROOT))
+
+from laiser.data_access import DataAccessLayer, KnowledgeFAISSIndexManager  # noqa: E402
+
+logging.basicConfig(level=logging.INFO, format="%(asctime)s [%(levelname)s] %(message)s")
+logger = logging.getLogger(__name__)
+
+# ---------------------------------------------------------------------------
+# Remote source URLs
+# ---------------------------------------------------------------------------
+# O*NET 28.0 individual files (plain-text tab-separated, UTF-8)
+ONET_KNOWLEDGE_URL = "https://www.onetcenter.org/dl_files/database/db_28_0_text/Knowledge.txt"
+ONET_CONTENT_MODEL_URL = "https://www.onetcenter.org/dl_files/database/db_28_0_text/Content%20Model%20Reference.txt"
+
+# ESCO skills CSV from LAiSER datasets repo (already used for skills index)
+ESCO_SKILLS_URL = (
+    "https://raw.githubusercontent.com/LAiSER-Software/datasets/refs/heads/master"
+    "/taxonomies/ESCO_skills_Taxonomy.csv"
+)
+
+# Minimum O*NET importance score to include (0–5 scale; 3.0 = "important")
+ONET_MIN_IMPORTANCE = 3.0
+
+OUTPUT_DIR = REPO_ROOT / "laiser" / "public"
+
+
+def _fetch(url: str, timeout: int = 60) -> bytes:
+    logger.info(f"Fetching {url} …")
+    resp = requests.get(url, timeout=timeout, verify=certifi.where())
+    resp.raise_for_status()
+    return resp.content
+
+
+def load_onet_knowledge(onet_dir: Path = None) -> pd.DataFrame:
+    """
+    Load O*NET Knowledge entries.
+
+    If onet_dir is provided, reads Knowledge.txt and Content Model Reference.txt
+    from that local directory.  Otherwise downloads them from onetcenter.org.
+
+    Returns DataFrame with columns: name, description, taxonomy
+    """
+    if onet_dir:
+        knowledge_path = onet_dir / "Knowledge.txt"
+        ref_path = onet_dir / "Content Model Reference.txt"
+        if not knowledge_path.exists():
+            raise FileNotFoundError(f"Knowledge.txt not found in {onet_dir}")
+        knowledge_raw = pd.read_csv(str(knowledge_path), sep="\t", dtype=str)
[…]
