+++ scripts/build_task_index.py
@@ -1,0 +1,252 @@
+"""
+Build the Task FAISS index for LAiSER v0.5.
+
+Sources:
+  1. O*NET Task Statements (Task Statements.txt)
+     Downloaded from: https://www.onetcenter.org/database.html#individual-files
+  2. ESCO Occupation Tasks filtered from the ESCO occupations CSV
+
+Output:
+  laiser/public/task_taxonomy.csv   — unified task taxonomy
+  laiser/public/tasks_v05.index     — FAISS IndexFlatIP
+  laiser/public/tasks_df.json       — metadata JSON
+
+Usage:
+  python scripts/build_task_index.py
+  python scripts/build_task_index.py --force-rebuild
+  python scripts/build_task_index.py --onet-dir /path/to/onet_db_28_0/
+"""
+
+import argparse
+import io
+import logging
+import sys
+from pathlib import Path
+
+import certifi
+import pandas as pd
+import requests
+
+REPO_ROOT = Path(__file__).resolve().parent.parent
+sys.path.insert(0, str(REPO_ROOT))
+
+from laiser.data_access import DataAccessLayer, TaskFAISSIndexManager  # noqa: E402
+
+logging.basicConfig(level=logging.INFO, format="%(asctime)s [%(levelname)s] %(message)s")
+logger = logging.getLogger(__name__)
+
+# ---------------------------------------------------------------------------
+# Remote source URLs
+# ---------------------------------------------------------------------------
+ONET_TASKS_URL = "https://www.onetcenter.org/dl_files/database/db_28_0_text/Task%20Statements.txt"
+ONET_OCCUPATION_URL = "https://www.onetcenter.org/dl_files/database/db_28_0_text/Occupation%20Data.txt"
+
+# ESCO occupations CSV — tasks are embedded as "essential" / "optional" skill statements
+# We use the ESCO skills CSV which includes occupational context via altLabels
+ESCO_OCCUPATIONS_URL = (
+    "https://raw.githubusercontent.com/LAiSER-Software/datasets/refs/heads/master"
+    "/taxonomies/ESCO_skills_Taxonomy.csv"
+)
+
+# Minimum O*NET incumbents responding % to include a task (0–100; 50 = majority reported it)
+ONET_MIN_INCUMBENTS_PCT = 50.0
+
+OUTPUT_DIR = REPO_ROOT / "laiser" / "public"
+
+
+def _fetch(url: str, timeout: int = 60) -> bytes:
+    logger.info(f"Fetching {url} …")
+    resp = requests.get(url, timeout=timeout, verify=certifi.where())
+    resp.raise_for_status()
+    return resp.content
+
+
+def _extract_action_verb(task_text: str) -> str:
+    """Return the first word of a task statement (the action verb)."""
+    if not task_text:
+        return ""
+    return task_text.strip().split()[0].rstrip(".,;:")
+
+
+def load_onet_tasks(onet_dir: Path = None) -> pd.DataFrame:
+    """
+    Load O*NET Task Statements.
+
+    Filters to Core tasks with >= ONET_MIN_INCUMBENTS_PCT incumbents responding.
+    Joins with Occupation Data for occupation title context.
+
+    Returns DataFrame with columns: name, description, taxonomy, action_verb
+    """
+    if onet_dir:
+        tasks_path = onet_dir / "Task Statements.txt"
+        occ_path = onet_dir / "Occupation Data.txt"
+        if not tasks_path.exists():
+            raise FileNotFoundError(f"Task Statements.txt not found in {onet_dir}")
+        tasks_raw = pd.read_csv(str(tasks_path), sep="\t", dtype=str)
+        occ_raw = pd.read_csv(str(occ_path), sep="\t", dtype=str) if occ_path.exists() else None
+    else:
[…]
