+++ .github/workflows/deploy-preview.yml
@@ -17,7 +17,7 @@ jobs:
       pull-requests: write   # required for PR comment
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
 
       - uses: actions/setup-python@v5
         with:
