+++ main.py
@@ -1,16 +1,31 @@
-import torch
-import pandas as pd
 import argparse
+
+import pandas as pd
+import torch
+
 from laiser.skill_extractor import Skill_Extractor
 
 # TODO: Check with and without GPU availablility
 
 # Parse command-line arguments
-parser = argparse.ArgumentParser(description='Run Skill Extractor on jobs and syllabi data.')
-parser.add_argument('--HF_TOKEN', type=str, default=None, help='Hugging Face token for authentication')
-parser.add_argument('--AI_MODEL_ID', type=str, default=None, help='Model name for Skill Extractor')
-parser.add_argument('--use_gpu', type=str, default=str(torch.cuda.is_available()), help='Enable or disable GPU use.')
-parser.add_argument('--batch_size', type=int, default=32, help='Batch size for skills extraction')
+parser = argparse.ArgumentParser(
+    description="Run Skill Extractor on jobs and syllabi data."
+)
+parser.add_argument(
+    "--HF_TOKEN", type=str, default=None, help="Hugging Face token for authentication"
+)
+parser.add_argument(
+    "--AI_MODEL_ID", type=str, default=None, help="Model name for Skill Extractor"
+)
+parser.add_argument(
+    "--use_gpu",
+    type=str,
+    default=str(torch.cuda.is_available()),
+    help="Enable or disable GPU use.",
+)
+parser.add_argument(
+    "--batch_size", type=int, default=32, help="Batch size for skills extraction"
+)
 args = parser.parse_args()
 
 
@@ -19,44 +34,58 @@ use_gpu = True if args.use_gpu == "True" else False
 # based on the above arguments, to run this script:
 # python main.py --HF_TOKEN <your_hf_token> --AI_MODEL_ID <your_model_id> --use_gpu True --batch_size 32
 
-print('\n\nInitializing the Skill Extractor...')
-se = Skill_Extractor(AI_MODEL_ID=args.AI_MODEL_ID, HF_TOKEN=args.HF_TOKEN, use_gpu=use_gpu)
-print('The Skill Extractor has been initialized successfully!\n')
+print("\n\nInitializing the Skill Extractor...")
+se = Skill_Extractor(
+    AI_MODEL_ID=args.AI_MODEL_ID, HF_TOKEN=args.HF_TOKEN, use_gpu=use_gpu
+)
+print("The Skill Extractor has been initialized successfully!\n")
 
 # Skill extraction from jobs data
-print('\n\nLoading a sample dataset of 50 jobs...')
-job_sample = pd.read_csv('https://raw.githubusercontent.com/LAiSER-Software/datasets/refs/heads/master/jobs-data/linkedin_jobs_sample_36rows.csv')
-print('The sample jobs dataset has been loaded successfully!\n')
+print("\n\nLoading a sample dataset of 50 jobs...")
+job_sample = pd.read_csv(
+    "https://raw.githubusercontent.com/LAiSER-Software/datasets/refs/heads/master/jobs-data/linkedin_jobs_sample_36rows.csv"
+)
+print("The sample jobs dataset has been loaded successfully!\n")
 
-job_sample = job_sample[['description', 'job_id']]
+job_sample = job_sample[["description", "job_id"]]
 job_sample = job_sample[1:3]
-print('The sample dataset has been filtered successfully!\n')
-print('Head of the sample:\n', job_sample.head())
+print("The sample dataset has been filtered successfully!\n")
+print("Head of the sample:\n", job_sample.head())
 
-output = se.extractor(job_sample, 'job_id', text_columns=['description'], batch_size=args.batch_size)
-print('The skills have been extracted from jobs data successfully...\n')
+output = se.extractor(
+    job_sample, "job_id", text_columns=["description"], batch_size=args.batch_size
+)
[…]
