### 2026-03-31 cef27db
fix: wire source_url into top_k zip and result dataframe
