+++ .pre-commit-config.yaml
@@ -14,4 +14,4 @@ repos:
     rev: 7.1.1
     hooks:
       - id: flake8
-        args: [--max-line-length=120, --extend-ignore=E203,W503,E266,W291,E501,E402]
+        args: ["--max-line-length=120", "--extend-ignore=E203,W503,E266,W291,E501,E402"]
