### 2026-04-07 #9336 @atrisovic
Hi @a9p @dpalmasan @jansim,

I hope you’re doing well! I’m reaching out to invite you to review a submission to the Journal of Open Source Software titled “LAiSER: A Taxonomy-Aware Framework for Labor Market Intelligence.”

Given your expertise in this area, I believe your feedback would be incredibly valuable in helping evaluate this work. The review would take place through JOSS’s open review process on GitHub. Please let me know if you would be willing to serve as a reviewer, or if you have any questions.

Thank you for your time and consideration.

Best,
Ana Trisovic,
Topic Editor
