### 2026-04-16 #10421 @editorialbot
[REVIEW]: LAiSER: A Taxonomy-Aware Framework for Labor Market Intelligence
**Submitting author:** @phanindra-max (<a href="http://orcid.org/0009-0004-6447-3760">Phanindra Kalaga</a>)
**Repository:** https://github.com/LAiSER-Software/extract-module
**Branch with paper.md** (empty if default branch): main
**Version:** v0.31
