### 2026-05-06 1655073
feat(gemini): update Gemini integration
