+++ pyproject.toml
@@ -37,7 +37,7 @@ dependencies = [
     "certifi>=2023.0.0",
     "requests>=2.25.0",
     "psutil",
-    "google-generativeai==0.8.5"
+    "google-genai>=0.8.0"
 ]
 
 [project.optional-dependencies]
