### 2026-05-06 52044af
feat(kst): add concept extraction API
