+++ laiser/skill_extractor_refactored.py
@@ -73,6 +73,32 @@ class SkillExtractorRefactored:
     to appropriate service classes.
     """
 
+    @staticmethod
+    def _expand_allowed_sources(allowed_sources: Optional[List[str]]) -> Optional[List[str]]:
+        if not allowed_sources:
+            return allowed_sources
+
+        expanded: List[str] = []
+        seen = set()
+        source_map = {
+            "esco": ["esco", "esco_knowledge", "esco_task"],
+            "onet": ["onet_skill", "onet_tech", "onet_knowledge", "onet_task"],
+            "osn": ["osn"],
+        }
+
+        for source in allowed_sources:
+            normalized = str(source or "").strip().lower()
+            if not normalized:
+                continue
+            mapped_values = source_map.get(normalized, [normalized])
+            for value in mapped_values:
+                if value in seen:
+                    continue
+                seen.add(value)
+                expanded.append(value)
+
+        return expanded
+
     def __init__(
         self,
         model_id: Optional[str] = None,
@@ -106,7 +132,7 @@ class SkillExtractorRefactored:
             backend=backend,
         )
 
-    def extract_and_align(
+    def extract_concepts(
         self,
         data: pd.DataFrame,
         id_column: str = "Research ID",
@@ -118,12 +144,14 @@ class SkillExtractorRefactored:
         batch_size: int = DEFAULT_BATCH_SIZE,
         warnings: bool = False,
         allowed_sources: Optional[List[str]] = None,
+        concepts: List[str] = None,
         extract: List[str] = None,
         return_edges: bool = False,
         similarity_thresholds: Optional[Dict[str, float]] = None,
+        timing: bool = False,
     ):
         """
-        Extract and align skills (and optionally Knowledge + Tasks) from a dataset.
+        Extract concepts from text and align them to taxonomy data.
 
         Parameters
         ----------
@@ -149,19 +177,32 @@ class SkillExtractorRefactored:
             Batch size for processing
         warnings : bool
             Whether to show warnings
-        extract : list, optional
+        concepts : list, optional
             Types to extract: "skills", "knowledge", "tasks", or ["all"].
-            Defaults to ["skills"] for backward compatibility.
+            Defaults to ["skills"].
+        extract : list, optional
+            Backward-compatible alias for `concepts`.
         return_edges : bool, optional
             If True, return {"nodes": pd.DataFrame, "edges": pd.DataFrame} where
             "edges" contains ENABLES edges (Knowledge → Task per skill).
             If False (default), return a plain pd.DataFrame.
+        timing : bool, optional
+            If True, print per-stage timing for extraction and alignment steps.
 
         Returns
         -------
         pd.DataFrame  (when return_edges=False)
+            Normalized mixed-concept rows with:
+            Research ID, Type, Raw Concept, Taxonomy Concept,
+            Taxonomy Description, Taxonomy Source, Source Url, Correlation Coefficient
         dict          (when return_edges=True): {"nodes": pd.DataFrame, "edges": pd.DataFrame}
         """
+        if concepts is not None and extract is not None and concepts != extract:
+            raise ValueError("Pass either `concepts` or `extract`, not both")
+
+        requested_concepts = concepts if concepts is not None else extract
+        expanded_sources = self._expand_allowed_sources(allowed_sources)
+
         return self.skill_service.extract_and_align_core(
             data=data,
             id_column=id_column,
@@ -172,11 +213,73 @@ class SkillExtractorRefactored:
             levels=levels,
             batch_size=batch_size,
             warnings=warnings,
-            allowed_sources=allowed_sources,
-            extract=extract,
+            allowed_sources=expanded_sources,
+            extract=requested_concepts,
             return_edges=return_edges,
[…]
