+++ laiser/services.py
@@ -261,7 +261,6 @@ class ResponseParser:
             return []
 
     @staticmethod
-    @staticmethod
     def parse_knowledge_task_response(response: str) -> List[Dict[str, Any]]:
         """
         Parse the JSON response from KT_FROM_SKILLS_PROMPT.
@@ -271,19 +270,23 @@ class ResponseParser:
             knowledge List[str]
             tasks     List[str]
         """
-        try:
-            # Strip markdown code fences if present
-            cleaned = re.sub(r"```(?:json)?", "", response).strip().rstrip("`").strip()
-            parsed = json.loads(cleaned)
-            results = parsed.get("results", [])
+        if not response or not response.strip():
+            return []
 
+        def validate(results: Any) -> List[Dict[str, Any]]:
             validated = []
+            if results is None:
+                return validated
+            if not isinstance(results, list):
+                results = [results]
+
             for item in results:
                 if not isinstance(item, dict):
                     continue
-                skill = str(item.get("skill", "")).strip()
-                knowledge = item.get("knowledge", [])
-                tasks = item.get("tasks", [])
+
+                skill = str(item.get("skill", item.get("Skill", ""))).strip()
+                knowledge = item.get("knowledge", item.get("Knowledge", item.get("Knowledge Required", [])))
+                tasks = item.get("tasks", item.get("task", item.get("Task Abilities", [])))
 
                 if not isinstance(knowledge, list):
                     knowledge = [str(knowledge)] if knowledge else []
@@ -294,12 +297,99 @@ class ResponseParser:
                     validated.append(
                         {
                             "skill": skill,
-                            "knowledge": [k.strip() for k in knowledge if k.strip()],
-                            "tasks": [t.strip() for t in tasks if t.strip()],
+                            "knowledge": [str(k).strip() for k in knowledge if str(k).strip()],
+                            "tasks": [str(t).strip() for t in tasks if str(t).strip()],
                         }
                     )
             return validated
 
+        def parse_array_fragment(fragment: str) -> List[str]:
+            fragment = fragment.strip()
+            if not fragment:
+                return []
+
+            try:
+                parsed = json.loads(fragment)
+                if isinstance(parsed, list):
+                    return [str(x).strip() for x in parsed if str(x).strip()]
+            except Exception:
+                pass
+
+            return [m.strip() for m in re.findall(r'"([^"]+)"', fragment) if m.strip()]
+
+        try:
+            candidates: List[str] = []
+            stripped = response.strip()
+
+            fenced = re.search(r"```(?:json)?\s*(.*?)\s*```", stripped, re.DOTALL)
+            if fenced:
+                candidates.append(fenced.group(1).strip())
+
+            obj_match = re.search(r"\{.*\}", stripped, re.DOTALL)
+            if obj_match:
+                candidates.append(obj_match.group(0).strip())
+
+            list_match = re.search(r"\[.*\]", stripped, re.DOTALL)
+            if list_match:
+                candidates.append(list_match.group(0).strip())
+
+            candidates.append(stripped)
+
+            seen = set()
+            for candidate in candidates:
+                if not candidate or candidate in seen:
+                    continue
+                seen.add(candidate)
+
+                try:
+                    parsed = json.loads(candidate)
+                except json.JSONDecodeError:
+                    continue
+
+                if isinstance(parsed, dict):
+                    for key in ("results", "data", "items"):
+                        validated = validate(parsed.get(key, []))
+                        if validated:
[…]
