### 2026-05-06 e821ad5
feat(knowledge): rebuild knowledge taxonomy index
