### 2026-05-06 0974726
feat(task): rebuild task taxonomy index
