+++ scripts/build_task_index.py
@@ -2,7 +2,7 @@
 Build the Task FAISS index for LAiSER v0.5.
 
 Sources:
-  1. O*NET Task Statements (Task Statements.txt)
+  1. O*NET Task Statements (Task Statements.xlsx)
      Downloaded from: https://www.onetcenter.org/database.html#individual-files
   2. ESCO Occupation Tasks filtered from the ESCO occupations CSV
 
@@ -38,8 +38,8 @@ logger = logging.getLogger(__name__)
 # ---------------------------------------------------------------------------
 # Remote source URLs
 # ---------------------------------------------------------------------------
-ONET_TASKS_URL = "https://www.onetcenter.org/dl_files/database/db_28_0_text/Task%20Statements.txt"
-ONET_OCCUPATION_URL = "https://www.onetcenter.org/dl_files/database/db_28_0_text/Occupation%20Data.txt"
+ONET_TASKS_URL = "https://www.onetcenter.org/dl_files/database/db_30_2_excel/Task%20Statements.xlsx"
+ONET_OCCUPATION_URL = "https://www.onetcenter.org/dl_files/database/db_30_2_excel/Occupation%20Data.xlsx"
 
 # ESCO occupations CSV — tasks are embedded as "essential" / "optional" skill statements
 # We use the ESCO skills CSV which includes occupational context via altLabels
@@ -77,28 +77,42 @@ def load_onet_tasks(onet_dir: Path = None) -> pd.DataFrame:
 
     Returns DataFrame with columns: name, description, taxonomy, action_verb
     """
+    preserve_all_rows = False
     if onet_dir:
-        tasks_path = onet_dir / "Task Statements.txt"
-        occ_path = onet_dir / "Occupation Data.txt"
-        if not tasks_path.exists():
-            raise FileNotFoundError(f"Task Statements.txt not found in {onet_dir}")
-        tasks_raw = pd.read_csv(str(tasks_path), sep="\t", dtype=str)
-        occ_raw = pd.read_csv(str(occ_path), sep="\t", dtype=str) if occ_path.exists() else None
+        tasks_xlsx_path = onet_dir / "Task Statements.xlsx"
+        tasks_txt_path = onet_dir / "Task Statements.txt"
+        occ_xlsx_path = onet_dir / "Occupation Data.xlsx"
+        occ_txt_path = onet_dir / "Occupation Data.txt"
+
+        if tasks_xlsx_path.exists():
+            tasks_raw = pd.read_excel(str(tasks_xlsx_path), dtype=str)
+            preserve_all_rows = True
+        elif tasks_txt_path.exists():
+            tasks_raw = pd.read_csv(str(tasks_txt_path), sep="\t", dtype=str)
+        else:
+            raise FileNotFoundError(f"Task Statements.xlsx or Task Statements.txt not found in {onet_dir}")
+
+        if occ_xlsx_path.exists():
+            occ_raw = pd.read_excel(str(occ_xlsx_path), dtype=str)
+        elif occ_txt_path.exists():
+            occ_raw = pd.read_csv(str(occ_txt_path), sep="\t", dtype=str)
+        else:
+            occ_raw = None
     else:
-        tasks_raw = pd.read_csv(io.StringIO(_fetch(ONET_TASKS_URL).decode("utf-8")), sep="\t", dtype=str)
+        tasks_raw = pd.read_excel(io.BytesIO(_fetch(ONET_TASKS_URL)), dtype=str)
         try:
-            occ_raw = pd.read_csv(io.StringIO(_fetch(ONET_OCCUPATION_URL).decode("utf-8")), sep="\t", dtype=str)
[…]
