### 2026-05-06 1709ed3
test(kst): update extraction coverage
