+++ pyproject.toml
@@ -63,12 +63,24 @@ Documentation = "https://github.com/LAiSER-Software/extract-module/blob/main/REA
 
 [tool.setuptools]
 packages = {find = {where = ["."], include = ["laiser*"]}}
+include-package-data = false
 
 [tool.setuptools.dynamic]
 version = {attr = "laiser.__version__"}
 
 [tool.setuptools.package-data]
-laiser = ["public/*"]
+laiser = [
+    "public/faiss_skills.csv",
+    "public/knowledge_df.json",
+    "public/knowledge_taxonomy.csv",
+    "public/knowledge_v05.index",
+    "public/skills_df.json",
+    "public/skills_v04.index",
+    "public/task_taxonomy.csv",
+    "public/tasks_df.json",
+    "public/tasks_v05.index",
+]
+
 
 [tool.black]
 target-version = ["py311"]
