### 2026-06-19 cf1bd4a
ci: add python compatibility matrix
