### 2026-06-19 72e0c43
docs: refresh installation and usage guide
