+++ README.md
@@ -11,10 +11,10 @@
 LAiSER is a tool that helps learners, educators and employers share trusted and mutually intelligible information about skills​.
 
 - [About](#about)
+- [Architecture](#architecture)
 - [Requirements](#requirements)
 - [Setup and Installation](#setup-and-installation)
 - [Usage](#usage)
-  - [Google Colab Setup](#google-colab-setup)
 - [Funding](#funding)
 - [Authors](#authors)
 - [Partners](#partners)
@@ -27,25 +27,48 @@ By leveraging state-of-the-art AI models, LAiSER automates the process of identi
 
 The tool emphasizes standardization and transparency, offering a common framework that bridges the communication gap between different stakeholders. With LAiSER, educators can better align their teaching methods with industry requirements, and employers can more effectively identify the competencies required for their teams. The result is a more efficient and strategic approach to skill development, benefiting the entire ecosystem.
 
+## Architecture
+
+LAiSER uses a four-stage extraction and alignment pipeline:
+
+1. Extraction
+   Input text is normalized by input type and passed through prompt construction and LLM inference to produce raw concept candidates.
+2. Parsing and deduplication
+   Model output is parsed into structured concepts and filtered through exact and semantic deduplication.
+3. Taxonomy alignment
+   Extracted concepts are matched against bundled taxonomy indexes using embedding-based similarity search and threshold filtering.
+4. Output normalization
+   Alignment results are converted into a unified tabular schema, with optional edge generation for graph-style outputs.
+
 ## Requirements
-- Python version >= Python 3.9. 
-- A GPU with atleast 15GB video memory is essential for running this tool on large datasets.
+- Python version `>=3.8`.
+- The package supports the current tested matrix through Python `3.13`.
+- A GPU is recommended for heavy local model workflows, but API-backed extraction can run CPU-only.
+- Provider-specific environment variables may be required depending on backend:
+  - `GEMINI_API_KEY` or `GOOGLE_API_KEY`
+  - `OPENAI_API_KEY`
 
 ## Setup and Installation
 
-- Install LAiSER using pip:
+- Install LAiSER from PyPI:
 
-  ### For GPU support (recommended if you have a CUDA-capable GPU):
   ```shell
-  pip install laiser[gpu]
+  pip install laiser
   ```
-  ### For CPU-only environments:
-  ```shell
-  pip install laiser[cpu]
-  ```
-  ### By default, torch and vllm GPU dependencies are included. Only when using the [cpu] extra will these GPU dependencies be excluded.
 
-**NOTE**: Python 3.9 or later, *preferably 3.12*, is expected to be installed on your system. If you don't have Python installed, you can download it from [here](https://www.python.org/downloads/).
+- Install with GPU extras:
+
+  ```shell
+  pip install "laiser[gpu]"
+  ```
+
+- Install development dependencies from source:
+
+  ```shell
+  pip install -e ".[dev]"
+  ```
+
+**NOTE**: Python 3.8 or later is required. Python 3.12 or 3.13 is recommended for current development and CI parity.
 
 You can check if your machine has a GPU available with:
 ```shell
@@ -54,41 +77,95 @@ python -c "import torch; print(torch.cuda.is_available())"
 
 ## Usage
 
-As of now LAiSER can be used a python package in Google Colab or a local machine with GPU access. The steps to setup the tool are as follows:
+LAiSER is used as a Python package. The recommended API is `SkillExtractorRefactored`.
 
[…]
