### 2026-06-19 776e612
fix: align extraction runtime behavior
