### 2026-06-19 202e0e2
docs: add API examples
