### 2026-06-22 6cb5e22
chore(deps): bump orhun/git-cliff-action from 3 to 4
Bumps [orhun/git-cliff-action](https://github.com/orhun/git-cliff-action) from 3 to 4.
- [Release notes](https://github.com/orhun/git-cliff-action/releases)
- [Commits](https://github.com/orhun/git-cliff-action/compare/v3...v4)

---
updated-dependencies:
- dependency-name: orhun/git-cliff-action
  dependency-version: '4'
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
