+++ .github/workflows/pr-automation.yml
@@ -46,7 +46,7 @@ jobs:
           fetch-depth: 0
 
       - name: Calculate and label review time
-        uses: actions/github-script@v7
+        uses: actions/github-script@v9
         with:
           script: |
             const { data: pr } = await github.rest.pulls.get({
