+++ .github/workflows/ci.yml
@@ -122,7 +122,7 @@ jobs:
           exit 0
 
       - name: Post code hygiene summary
-        uses: actions/github-script@v7
+        uses: actions/github-script@v9
         with:
           script: |
             const found = '${{ steps.vulture.outputs.found }}' === 'true';
@@ -194,7 +194,7 @@ jobs:
           exit 0
 
       - name: Post complexity insights
-        uses: actions/github-script@v7
+        uses: actions/github-script@v9
         with:
           script: |
             const output = `${{ steps.radon.outputs.output }}`.trim();
