+++ .github/workflows/deploy-preview.yml
@@ -62,7 +62,7 @@ jobs:
           skip-existing: true
 
       - name: Comment on PR
-        uses: actions/github-script@v7
+        uses: actions/github-script@v9
         with:
           script: |
             github.rest.issues.createComment({
