+++ .github/workflows/draft-pdf.yml
@@ -8,7 +8,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@v7
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
