+++ .github/workflows/deploy-prod.yml
@@ -17,7 +17,7 @@ jobs:
       id-token: write  # required for trusted publishing
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v7
 
       - uses: actions/setup-python@v5
         with:
