+++ laiser/data_access.py
@@ -1,833 +1,0 @@
-"""
-Module Description:
--------------------
-This module provides the data access layer for the LAiSER project, handling all data loading, external API calls, and FAISS index management for skill extraction and research.
-
-Ownership:
-----------
-Project: Leveraging Artificial intelligence for Skills Extraction and Research (LAiSER)
-Owner:  George Washington University Insitute of Public Policy
-    Program on Skills, Credentials and Workforce Policy
-    Media and Public Affairs Building
-    805 21st Street NW
-    Washington, DC 20052
-    PSCWP@gwu.edu
-    https://gwipp.gwu.edu/program-skills-credentials-workforce-policy-pscwp
-
-License:
---------
-Copyright 2025 George Washington University Insitute of Public Policy
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files
-(the “Software”), to deal in the Software without restriction, including without limitation the rights to use, copy, modify,
-merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
-IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-
-Input Requirements:
--------------------
-- Requires CSV files for ESCO and combined skills taxonomies.
-- Requires access to remote FAISS index files and skill embedding models.
-
-Output/Return Format:
----------------------
-- Returns pandas DataFrames for skills data.
-- Produces and manages FAISS index files for skill similarity search.
-
-"""
-
-"""
-Revision History:
------------------
-Rev No.     Date            Author              Description
-[1.0.0]     08/10/2025      Satya Phanindra K.            Initial Version
-
-
-TODO:
------
-- 1: Add more robust error handling for data loading.
-- 2: Implement caching for remote data sources.
-"""
-
-import io
-import json
-import logging
-import os
-from pathlib import Path
-from typing import Any, Dict, List, Optional
-
-import certifi
-import faiss
-import numpy as np
-import pandas as pd
-import requests
-from sentence_transformers import SentenceTransformer
-
-from laiser.config import COMBINED_SKILLS_URL, DEFAULT_EMBEDDING_MODEL, ESCO_SKILLS_URL, OSN_SKILLS_URL
-from laiser.exceptions import FAISSIndexError, LAiSERError
-
-logger = logging.getLogger(__name__)
-
-
-class DataAccessLayer:
-    """Handles data loading and external API calls"""
-
-    def __init__(self):
-        self.embedding_model = None
-        self._esco_df = None
-        self._osn_df = None
-
-        self._combined_df = None
-
-    def get_embedding_model(self) -> SentenceTransformer:
-        """Get or initialize the embedding model"""
-        if self.embedding_model is None:
-            self.embedding_model = SentenceTransformer(DEFAULT_EMBEDDING_MODEL)
[…]
