### 2026-06-27 505ff40
test(extractor): refresh split module coverage
