+++ pytest.ini
@@ -6,3 +6,7 @@ markers =
     llm_cloud: cloud llms test
     openai: openai
     anthropic: anthropic
+filterwarnings =
+    ignore:builtin type SwigPyPacked has no __module__:DeprecationWarning
+    ignore:builtin type SwigPyObject has no __module__:DeprecationWarning
+    ignore:builtin type swigvarlink has no __module__:DeprecationWarning

[…]
