### 2026-06-27 1a578f4
docs(package): update references and guides
