### 2026-07-28 cca5bdd
fix: default all LLM backends to deterministic greedy decoding
Every backend previously inherited its provider's default temperature.
Set temperature 0.0 and wire GENERATION_SEED through the router to all
backends that accept one. Expose temperature and seed on
SkillExtractorRefactored. Add tests/test_determinism.py covering decoding
defaults, request payloads, router forwarding, and alignment stability.
Correct the paper's FAISS description from approximate to exact and add a
reference configuration to the README.

Addresses JOSS review issue #424.
