### 2026-08-04 1402dc4
Revert "fix: default all LLM backends to deterministic greedy decoding"
