+++ paper.md
@@ -45,13 +45,9 @@ LAiSER differs from these tools in several key respects. First, it uses LLMs for
 
 LAiSER's architecture is organized into three loosely coupled layers: data access, skill extraction, and taxonomy alignment. This separation was chosen to allow each component to evolve independently and to support diverse deployment scenarios without requiring changes to the overall pipeline.
 
-The extraction layer posed the most significant design trade-off. Training a custom NER model, as Nesta's library does, would yield fast inference but would require substantial labeled data for each new domain (syllabi, credentials, job postings). Instead, LAiSER delegates extraction to LLMs via a prompt engineering approach, trading some inference speed for broad domain generalizability without retraining. To mitigate vendor lock-in and hardware constraints, the extraction layer abstracts over multiple backends: vLLM for GPU clusters, HuggingFace Transformers and llama.cpp for local models, and the Gemini and OpenAI APIs for cloud-based inference. The active backend is resolved at initialization from the available hardware and credentials, so the same extraction code runs unchanged on a GPU cluster, a laptop, or a hosted API.
+The extraction layer posed the most significant design trade-off. Training a custom NER model, as Nesta's library does, would yield fast inference but would require substantial labeled data for each new domain (syllabi, credentials, job postings). Instead, LAiSER delegates extraction to LLMs via a prompt engineering approach, trading some inference speed for broad domain generalizability without retraining. To mitigate vendor lock-in and hardware constraints, the extraction layer abstracts over multiple backends (vLLM for GPU clusters, HuggingFace Transformers for local models, and the Gemini API for cloud-based inference), with automatic fallback when a backend is unavailable.
 
-For taxonomy alignment, the framework embeds both the extracted phrases and the taxonomy entries with a sentence-transformer model [@Reimers2019] and retrieves matches from a precomputed FAISS index [@Johnson2019]. The index is a flat inner-product index over L2-normalized embeddings, which makes retrieval exhaustive and exact rather than approximate. Exactness was a deliberate choice over an approximate index: at the scale of ESCO's 13,000+ skill entries an exhaustive scan is inexpensive relative to the cost of LLM inference in the preceding stage, and it guarantees that a given query returns the same neighbor and the same similarity score on every run, which is a precondition for the reproducibility properties described below. Because retrieval is encapsulated behind an index manager, an approximate index can be substituted without changes to the alignment logic should taxonomy size later make exhaustive search impractical. The system accepts pandas DataFrames as input and produces structured output including raw extracted skills, taxonomy-aligned canonical skills, taxonomy identifiers (e.g., ESCO codes), semantic similarity scores, and optional SCQF proficiency levels.
-
[…]
