+++ pytest.ini
@@ -6,4 +6,3 @@ markers =
     llm_cloud: cloud llms test
     openai: openai
     anthropic: anthropic
-    determinism: decoding determinism and alignment stability
