### 2026-08-04 a649ea2
fix(llm): route local generation to the backend that was actually loaded
LLMRouter.generate() sent every non-Gemini, non-OpenAI, non-llama.cpp
request to llm_generate_vllm(), regardless of what _initialize_components()
had loaded. On CPU, and whenever vLLM initialization fails, initialization
falls back to a Transformers tokenizer and model and leaves self.llm as
None, so generation passed None into the vLLM backend and failed. The
deterministic decoding defaults added to llm_generate() in #426 were
therefore correct but dormant, and paper.md could not claim a working
fallback.

Local dispatch
[…]
