+++ tests/test_determinism.py
@@ -18,11 +18,17 @@ or a paid API call:
 
 import importlib
 import inspect
+import os
 
 import pandas as pd
 import pytest
 
 from laiser.config import DEFAULT_TEMPERATURE, DEFAULT_TOP_P, GENERATION_SEED
+
+# pytest.ini registers a `determinism` marker; apply it to the whole module so
+# these tests can be selected or excluded as a group with `-m determinism`.
+# Individual alignment tests carry the `alignment` marker in addition.
+pytestmark = pytest.mark.determinism
 
 
 def _import_or_skip(module_path):
@@ -47,10 +53,30 @@ def _import_or_skip(module_path):
 
 
 def test_config_defaults_are_deterministic():
-    """The shipped configuration must select greedy decoding."""
-    assert DEFAULT_TEMPERATURE == 0.0, "Default decoding must be greedy (temperature 0.0)"
-    assert DEFAULT_TOP_P == 1.0, "Nucleus filtering must be inactive by default"
-    assert GENERATION_SEED is not None, "A default generation seed must be defined"
+    """The shipped configuration must select greedy decoding.
+
+    ``laiser.config`` reads LAISER_TEMPERATURE and LAISER_TOP_P at import time,
+    so asserting against the already-imported constants would assert whatever
+    the test runner's environment happens to export. A runner with
+    LAISER_TEMPERATURE=0.7 set would turn this guarantee into a false pass.
+    The module is therefore reloaded with those overrides removed, so that what
+    is asserted is the value LAiSER ships rather than the value in force.
+    """
+    import laiser.config as laiser_config
+
+    overrides = {
+        name: os.environ.pop(name)
+        for name in ("LAISER_TEMPERATURE", "LAISER_TOP_P")
+        if name in os.environ
+    }
+    try:
+        shipped = importlib.reload(laiser_config)
+        assert shipped.DEFAULT_TEMPERATURE == 0.0, "Default decoding must be greedy (temperature 0.0)"
+        assert shipped.DEFAULT_TOP_P == 1.0, "Nucleus filtering must be inactive by default"
+        assert shipped.GENERATION_SEED is not None, "A default generation seed must be defined"
+    finally:
+        os.environ.update(overrides)
+        importlib.reload(laiser_config)
 
 
 @pytest.mark.parametrize(
@@ -155,6 +181,17 @@ def test_anthropic_payload_carries_deterministic_decoding(monkeypatch):
     assert captured["temperature"] == DEFAULT_TEMPERATURE
 
 
+def _uninitialized_router(router_module, monkeypatch, **kwargs):
+    """Build an LLMRouter without loading a model.
+
+    ``_initialize_components`` downloads weights and starts an engine, so it is
+    stubbed out. Tests then set ``llm``/``model``/``tokenizer`` themselves to
+    describe the state a real initialization would have left behind.
+    """
+    monkeypatch.setattr(router_module.LLMRouter, "_initialize_components", lambda self: None)
+    return router_module.LLMRouter(model_id="local-model", use_gpu=False, **kwargs)
+
+
 def test_router_forwards_deterministic_decoding(monkeypatch):
     """LLMRouter must pass its temperature and seed through to the backend."""
     router_module = _import_or_skip("laiser.llm_models.llm_router")
@@ -167,13 +204,91 @@ def test_router_forwards_deterministic_decoding(monkeypatch):
         return "ok"
 
     monkeypatch.setattr(router_module, "llm_generate_vllm", _fake_vllm)
-    monkeypatch.setattr(router_module.LLMRouter, "_initialize_components", lambda self: None)
 
-    router = router_module.LLMRouter(model_id="local-model", use_gpu=False)
+    router = _uninitialized_router(router_module, monkeypatch)
[…]
