### 2026-08-04 c880fff
docs(paper): make the summary accessible and align backend claims with the code
Addresses the two parts of JOSS review issue #424 that PR #426 left
untouched, and corrects a backend description that the code has since
outgrown.

Summary. The reviewer noted it was too technical to be understood by a
reader without a technical background. It opened on a two-stage LLM and
FAISS pipeline and closed with pip invocations. It now opens on the problem
in plain language, that the same skill is described differently in a job
posting, a syllabus and a credential, which defeats analysis at scale;
explains
[…]
