+++ paper.md
@@ -29,7 +29,11 @@ bibliography: paper.bib
 
 # Summary
 
-LAiSER is an open-source Python framework for extracting, standardizing, and aligning skill information from unstructured text with established skill taxonomies. The system addresses the lack of standardized, machine-readable skill information needed to facilitate communication between learners, educators, and employers. LAiSER employs a two-stage pipeline combining large language models (LLMs) with semantic vector search using FAISS [@Johnson2019] for high-precision skill extraction and taxonomy alignment to standards such as ESCO [@ESCO2020]. The framework supports multiple computational backends, including GPU-accelerated inference, cloud APIs, and CPU-only environments, making it accessible across diverse research and operational contexts. LAiSER also supports extraction within the Knowledge, Skills, and Abilities (KSA) framework with automatic proficiency level classification using the Scottish Credit and Qualifications Framework (SCQF) [@SCQF2019]. The source code is available at [https://github.com/LAiSER-Software/extract-module](https://github.com/LAiSER-Software/extract-module) under the MIT License, installable via `pip install laiser[gpu]` or `pip install laiser[cpu]`.
+Job postings, course syllabi, and credential descriptions all describe skills, but each describes them in its own words. One employer asks for "proficiency in Python," another for "experience building data pipelines," and a college course promises "programming fundamentals." A person reading these can see which ones overlap; software cannot, and neither can the large-scale analyses that educators, employers, and policymakers rely on to judge whether what is taught matches what is hired for.
+
+LAiSER is an open-source Python tool that reads text of this kind, uses large language models (LLMs) to identify the skills it describes, and matches each one to an entry in a published catalogue of skills. Such catalogues, known as skill taxonomies, give every skill a standard name and a stable identifier, so that the three phrasings above resolve to the same entry and become countable and comparable across thousands of documents. LAiSER aligns to the European Skills, Competences, Qualifications and Occupations classification (ESCO) [@ESCO2020] among other taxonomies, and can additionally sort what it finds into knowledge, skills, and abilities (KSA) and estimate the level at which each is required, using the Scottish Credit and Qualifications Framework (SCQF) [@SCQF2019].
+
+The intended users are researchers and practitioners who need to analyse skill language at a scale that makes reading by hand impractical: labor economists studying how occupations change, curriculum designers checking a program against industry demand, and credentialing bodies issuing standards-aligned badges. LAiSER runs on a laptop, on a shared GPU cluster, or against a hosted model service, so the same analysis is available whether or not a research group has dedicated hardware. The source code is available at [https://github.com/LAiSER-Software/extract-module](https://github.com/LAiSER-Software/extract-module) under the MIT License.
 
 # Statement of need
 
@@ -37,17 +41,17 @@ The contemporary labor market is characterized by rapid technological change, ev
 
 # State of the field
 
-Several open-source tools address parts of the skill extraction problem. SkillNER [@SkillNER2022] uses rule-based matching against the EMSI skills database via spaCy, offering fast lookups but limited ability to detect novel or context-dependent skills not present in its dictionary. Nesta's Skills Extractor Library [@Nesta2023] trains a spaCy NER model on labeled job advertisements and maps extracted phrases to ESCO or Lightcast taxonomies using sentence-transformer embeddings. While effective for UK job advertisements, its NER model requires domain-specific labeled training data and does not generalize well to other text types such as syllabi or credential descriptions. The esco-skill-extractor [@EscoExtractor2024] takes a simpler approach, embedding full input texts with a transformer and matching against ESCO entries by cosine similarity, but it lacks preprocessing, multi-model flexibility, and proficiency-level classification.
[…]
