+++ paper.md
@@ -45,7 +45,7 @@ LAiSER differs from these tools in several key respects. First, it uses LLMs for
 
 LAiSER's architecture is organized into three loosely coupled layers: data access, skill extraction, and taxonomy alignment. This separation was chosen to allow each component to evolve independently and to support diverse deployment scenarios without requiring changes to the overall pipeline.
 
-The extraction layer posed the most significant design trade-off. Training a custom NER model, as Nesta's library does, would yield fast inference but would require substantial labeled data for each new domain (syllabi, credentials, job postings). Instead, LAiSER delegates extraction to LLMs via a prompt engineering approach, trading some inference speed for broad domain generalizability without retraining. To mitigate vendor lock-in and hardware constraints, the extraction layer abstracts over multiple backends (vLLM for GPU clusters, HuggingFace Transformers for local models, and the Gemini API for cloud-based inference), with automatic fallback when a backend is unavailable.
+The extraction layer posed the most significant design trade-off. Training a custom NER model, as Nesta's library does, would yield fast inference but would require substantial labeled data for each new domain (syllabi, credentials, job postings). Instead, LAiSER delegates extraction to LLMs via a prompt engineering approach, trading some inference speed for broad domain generalizability without retraining. To mitigate vendor lock-in and hardware constraints, the extraction layer abstracts over multiple backends (vLLM for GPU clusters, HuggingFace Transformers for local models, llama.cpp for local GGUF models, and the Gemini and OpenAI APIs for cloud-based inference), with automatic fallback when a backend is unavailable.
 
 For taxonomy alignment, the framework uses FAISS for approximate nearest-neighbor search over sentence-transformer embeddings [@Reimers2019] rather than brute-force cosine similarity. This choice enables sub-linear scaling as taxonomy size grows, which is essential when aligning against ESCO's 13,000+ skill entries at production volumes. The system accepts pandas DataFrames as input and produces structured output including raw extracted skills, taxonomy-aligned canonical skills, taxonomy identifiers (e.g., ESCO codes), semantic similarity scores, and optional SCQF proficiency levels.
 
