### 2026-08-04 03405ef
fix: kwargs for gemini models
Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>
