### 2026-08-19 1c88af9
Revert "fix: default all LLM backends to deterministic greedy decoding" (#428)
* Revert "fix: default all LLM backends to deterministic greedy decoding"

* Potential fix for pull request finding

Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>

* Potential fix for pull request finding

Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>

* fix: kwargs for gemini models

Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot Autofix powered by
[…]
