+++ laiser/llm_models/hugging_face_llm.py
@@ -49,10 +49,7 @@ Rev No.     Date            Author              Description
 [1.0.0]     6/30/2025      Anket Patil          Modularize LLM generation logic for transformers and vLLM
 """
 
-import torch
 from transformers import AutoModelForCausalLM, AutoTokenizer, BitsAndBytesConfig
-
-from laiser.config import DEFAULT_TEMPERATURE, DEFAULT_TOP_P, GENERATION_SEED
 
 try:
     from vllm import SamplingParams
@@ -63,24 +60,7 @@ except ImportError:
     SamplingParams = None  # Optional fallback
 
 
-def llm_generate(
-    prompt: str,
-    tokenizer,
-    model,
-    model_id: str,
-    use_gpu: bool,
-    temperature: float = DEFAULT_TEMPERATURE,
-    seed: int = GENERATION_SEED,
-):
-    """Generate text with a local HuggingFace Transformers model.
-
-    Decoding is deterministic by default. A temperature of 0.0 disables
-    sampling entirely (``do_sample=False``, i.e. greedy decoding), which also
-    prevents a checkpoint's bundled ``generation_config.json`` from silently
-    re-enabling sampling. When a caller raises the temperature above zero,
-    ``seed`` is applied through ``torch.manual_seed`` so the run stays
-    reproducible.
-    """
+def llm_generate(prompt: str, tokenizer, model, model_id: str, use_gpu: bool):
     if tokenizer is None or model is None:
         quantization_config = BitsAndBytesConfig(load_in_8bit=True)
 
@@ -91,50 +71,23 @@ def llm_generate(
 
     inputs = tokenizer(prompt, return_tensors="pt").to(model.device)
 
-    do_sample = temperature is not None and temperature > 0.0
-    if seed is not None:
-        torch.manual_seed(seed)
-
-    generation_kwargs = {
-        "max_new_tokens": 100,
-        "pad_token_id": tokenizer.pad_token_id,
-        "eos_token_id": tokenizer.eos_token_id,
-        "do_sample": do_sample,
-    }
-    if do_sample:
-        generation_kwargs["temperature"] = temperature
-        generation_kwargs["top_p"] = DEFAULT_TOP_P
-
-    outputs = model.generate(**inputs, **generation_kwargs)
+    outputs = model.generate(
+        **inputs,
+        max_new_tokens=100,
+        pad_token_id=tokenizer.pad_token_id,
+        eos_token_id=tokenizer.eos_token_id,
+    )
 
     return tokenizer.decode(outputs[0], skip_special_tokens=True)
 
 
-def llm_generate_vllm(
-    prompt,
-    llm,
-    temperature: float = DEFAULT_TEMPERATURE,
-    seed: int = GENERATION_SEED,
-    max_tokens: int = 200,
-):
-    """Generate text with a local vLLM engine.
-
-    Decoding is deterministic by default: ``temperature`` defaults to
-    ``config.DEFAULT_TEMPERATURE`` (0.0, i.e. greedy decoding) rather than to
-    vLLM's own default of 1.0, and ``config.GENERATION_SEED`` is passed so that
-    runs remain reproducible if a caller raises the temperature.
-    """
+def llm_generate_vllm(prompt, llm):
     if not VLLM_AVAILABLE:
         raise ImportError(
             "vLLM is not installed. Please install it to use this function."
         )
 
-    sampling_params = SamplingParams(
-        max_tokens=max_tokens,
-        temperature=temperature,
-        top_p=DEFAULT_TOP_P,
-        seed=seed,
-    )
+    sampling_params = SamplingParams(max_tokens=200, seed=42)
     result = llm.generate([prompt], sampling_params=sampling_params)
     raw_text = result[0].outputs[0].text.strip()
     return raw_text
