### 2026-08-19 81987b1
test(kst): update extraction sample
