### 2026-08-19 f11eeac
fix(kst): add response diagnostics
