+++ laiser/services.py
@@ -1003,6 +1003,14 @@ class SkillExtractionService:
         response = self.router.generate(prompt)
         results = self.llm_parser.parse_knowledge_task_response(response)
 
+        logger.warning(
+            "KT model response: %d parsed blocks, %d chars; preview=%s; tail=%s",
+            len(results),
+            len(response),
+            response.strip().replace("\n", " ")[:500],
+            response.strip().replace("\n", " ")[-500:],
+        )
+
         if not results:
             preview = response.strip().replace("\n", " ")[:200]
             logger.warning(f"Failed to parse knowledge/task response: {preview}")
