### 2026-08-19 fdf7769
docs(readme): update funding and taxonomy examples
