### 2026-08-25 384df20
fix(runtime): limit OpenMP threads on import
