+++ laiser/__init__.py
@@ -4,6 +4,8 @@ LAiSER - Leveraging Artificial Intelligence for Skills Extraction and Research
 A Python package for extracting and aligning skills from text using AI models.
 """
 
+import os
+os.environ.setdefault("OMP_NUM_THREADS", "1")
 __version__ = "0.5"
 
 # Import main classes for easy access
