+++ .github/workflows/draft-pdf.yml
@@ -4,6 +4,7 @@ on: [push]
 
 jobs:
   paper:
+    if: ${{ false }}
     runs-on: ubuntu-latest
     name: Paper Draft
     steps:
@@ -23,4 +24,3 @@ jobs:
           # PDF. Note, this should be the same directory as the input
           # paper.md
           path: paper.pdf
-
