+++ .github/workflows/codeql.yml
@@ -9,6 +9,7 @@ on:
 
 jobs:
   analyze:
+    if: ${{ false }}
     name: CodeQL Analysis
     runs-on: ubuntu-latest
     permissions:
