### 2026-08-25 a65e880
Revert "chore(deps-dev): update python-dotenv requirement from <1.0.0,>=0.19.0 to >=0.19.0,<2.0.0"
