+++ pyproject.toml
@@ -52,7 +52,7 @@ dev = [
     "flake8>=4.0",
     "build>=0.8.0",
     "twine>=4.0.0",
-    "python-dotenv>=0.19.0,<2.0.0",
+    "python-dotenv>=0.19.0,<1.0.0",
 ]
 
 [project.urls]
