### 2026-09-02 LAiSER-Software/extract-module#437 @jansim
[JOSS] Problems with CI
I think there are a few issues regarding the CI/CD setup in the repo. Grouping these into a single issue to avoid clutter, but feel free to spread out into multiple separate issues.

- [ ] It seems that the actual test battery is not being run as part of the CI/CD pipeline and only a basic compatibility test of importing the package is run? I think it makes sense to not run parts of the test suite that would use API tokens, but those could be skipped in the CI environment and it would be good to run as many tests as possible also in CI.
- [ ] The CI pipeline is currently only executed on pull_request. You might want to add another hook for on push (maybe only for the main branch).
- [ ] On recent commits on the main branch the CI pipeline is broken
- [ ] One usually wants to avoid the chance of breaking the CI pipeline on main. It could make sense here to set up branch protections, such that PRs can only be merged if the checks pass.

I'm trying to be concise, so if something is unclear here please don't hesitate to ask and happy to discuss if you disagree with any of these suggestions.

Related to [open-journals/joss-reviews#10421](https://github.com/openjournals/joss-reviews/issues/10421).
