### 2026-09-02 LAiSER-Software/extract-module#438 @jansim
[JOSS] Issues / suggestions documentation
- [ ] Note blocks on https://laiser-software.github.io/docs/installation/#ii-importing-skill-extractor-from-laiser-package seem to be ill-formatted
- [ ] Usage page needs some updates to the prose and fixes to links https://laiser-software.github.io/docs/usage/. There's also a lack of actual explanations and mostly screenshots following onto one another.
- [ ] It seems the documentation website in general hasn't been examined in a while and requires some more work / clean-up. It would be good to match the information there to the one from e.g. the repo README.
- [ ] If Google Colab is the preferred mode of execution, maybe provide a link to quick start notebook on the README too. Or is this information outdated?
- [ ] If possible, maybe the usage sections might also include example code that does not require an API key?
- [ ] Is there documentation somewhere on e.g. which model providers are supported or how the functions are to be used in more detail? It seems like mostly quick "Getting started" / set-up docs.

Related to https://github.com/openjournals/joss-reviews/issues/10421.
