### 2026-09-02 #10421 @jansim
Pausing my review for now as it would be greatly aided by better documentation of the software. It would also be great to know whether there's a way of using the tool without requiring an API key? (It's ok if that leads to significantly worse results, it's less about evaluating performance and mostly just overall function)
