### 2026-09-07 31dd8b3
Add CITATION.cff
