### 2026-09-10 758ea61
fix (CI): outdated Docker image that causes failure
