### 2026-09-11 a3cc7ce
style: apply black formatting
The CI lint job runs `black --check .`, which failed on 13 files. Running
black resolves all of them, and incidentally clears the three flake8 errors
that were also outstanding (E302 in llm_models/gemini.py, E303 in
llm_methods.py and utils.py).

Formatting only, no behaviour change: the test suite is unchanged at
3 passed, 5 skipped.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
