+++ laiser/llm_models/llama_cpp_handler.py
@@ -20,7 +20,6 @@ def _strip_fences(text: str) -> str:
 
 
 class LlamaCppBackend:
-
     def __init__(
         self,
         model_path: Optional[str] = None,
@@ -35,9 +34,7 @@ class LlamaCppBackend:
 
         model_path = model_path or os.getenv("LAISER_LLAMA_CPP_MODEL_PATH")
         if not model_path:
-            raise ValueError(
-                "Set LAISER_LLAMA_CPP_MODEL_PATH or pass model_path to LlamaCppBackend."
-            )
+            raise ValueError("Set LAISER_LLAMA_CPP_MODEL_PATH or pass model_path to LlamaCppBackend.")
 
         model_path = str(Path(model_path))
         # model_path = str(Path(model_path).expanduser().resolve())
@@ -107,11 +104,8 @@ def llama_cpp_chat(
     stop: Optional[List[str]] = None,
     temperature: Optional[float] = None,
 ) -> str:
-
     if llama is None:
-        raise ValueError(
-            "llama is None; expected an initialized llama_cpp.Llama instance."
-        )
+        raise ValueError("llama is None; expected an initialized llama_cpp.Llama instance.")
 
     messages = [
         {"role": "system", "content": system},
