### 2026-09-11 14c6e85
fix(ci): repair release and PR workflows, run the real test battery
Addresses the CI problems raised in #437.

Publishing:
- deploy-prod.yml now triggers on v* tags instead of every push to main.
  Pushing to main re-uploaded an already-published version, which PyPI
  rejects with "File already exists", so the job could only ever succeed
  once per version bump.
- Delete python-publish.yml. It was still the unedited GitHub template and
  was a second publisher for the same project; since PyPI trusted publishing
  is scoped to a workflow filename, only one of the two could authenticate.
- Guard
[…]
