+++ .github/workflows/ci.yml
@@ -2,11 +2,14 @@ name: CI
 
 on:
   pull_request:
+  push:
+    branches:
+      - main
+      - dev
 
 jobs:
   # ── 1. LINT ────────────────────────────────────────────────────────────────
   lint:
-    if: ${{ false }}
     name: Lint
     runs-on: ubuntu-latest
     steps:
@@ -29,9 +32,37 @@ jobs:
         run: isort --check-only .
 
 
+  # ── 2. TEST ────────────────────────────────────────────────────────────────
+  test:
+    name: Test
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Cache pip
+        uses: actions/cache@v4
+        with:
+          path: ~/.cache/pip
+          key: ${{ runner.os }}-pip-test-${{ hashFiles('pyproject.toml') }}
+          restore-keys: |
+            ${{ runner.os }}-pip-test-
+
+      - name: Install package and test dependencies
+        run: pip install -e ".[dev]"
+
+      - name: pytest
+        # No secrets are supplied here on purpose. Tests that need provider
+        # credentials skip themselves when the key is absent (see the guards in
+        # tests/test_openai.py, test_anthropic.py, test_kst_extraction.py and
+        # test_llm_cloud.py), so the battery runs green without any token.
+        run: pytest tests/ -v -rs --cov=laiser --cov-report=term-missing
+
   # ── 3. SECURITY ────────────────────────────────────────────────────────────
   security:
-    if: ${{ false }}
     name: Security
     runs-on: ubuntu-latest
     steps:
@@ -59,9 +90,8 @@ jobs:
         # CVE-2026-4539 ignored - affects pygments 2.19.2, no fix available yet
         run: pip-audit --ignore-vuln CVE-2026-4539
 
-  # ── 3. TYPE CHECK ──────────────────────────────────────────────────────────
+  # ── 4. TYPE CHECK ──────────────────────────────────────────────────────────
   type-check:
-    if: ${{ false }}
     name: Type Check
     runs-on: ubuntu-latest
     steps:
@@ -78,9 +108,8 @@ jobs:
         run: mypy laiser/ --ignore-missing-imports --no-strict-optional
         continue-on-error: true
 
-  # ── 4. LICENSE CHECK ───────────────────────────────────────────────────────
+  # ── 5. LICENSE CHECK ───────────────────────────────────────────────────────
   license-check:
-    if: ${{ false }}
     name: License Check
     runs-on: ubuntu-latest
     steps:
@@ -97,10 +126,11 @@ jobs:
         # Fail if any GPL licensed dependency is found
         run: pip-licenses --fail-on="GNU General Public License v2 (GPLv2);GNU General Public License v3 (GPLv3);GNU Affero General Public License v3"
 
-  # ── 5. DEAD CODE ───────────────────────────────────────────────────────────
+  # ── 6. DEAD CODE ───────────────────────────────────────────────────────────
   dead-code:
-    if: ${{ false }}
     name: Dead Code
+    # Reports by commenting on the PR, so there is nothing to post to on a push.
+    if: github.event_name == 'pull_request'
     runs-on: ubuntu-latest
     permissions:
       pull-requests: write
@@ -128,10 +158,16 @@ jobs:
 
       - name: Post code hygiene summary
         uses: actions/github-script@v7
+        env:
+          # Passed via env, never interpolated into the script body: vulture
+          # output contains repo file paths, which are attacker-controlled on
+          # a pull_request and would otherwise break out of the template literal.
+          VULTURE_FOUND: ${{ steps.vulture.outputs.found }}
[…]
