### 2026-09-12 20fca78
fix(ci): resolve the failures the first pipeline run surfaced
Enabling the workflows made four jobs run for the first time. All four
failed; this addresses them.

Test — tests/test_index_process.py asserted that skill_embeddings.npy exists
after a *cached* index load. That file is a build product ignored by
.gitignore:191 (commit 26f4c62), so it is absent from any fresh checkout and
the cache path does not generate one. Step 1 now checks only the committed
artifacts; steps 3 and 5 still expect all three, since those force a rebuild.
The assertion was wrong everywhere, not just on CI — the arti
[…]
