### 2026-09-12 85ff07e
fix: address Copilot review findings on #440
Taxonomy source aliases (laiser/skill_extractor_refactored.py)
f57e1b6 relabelled the taxonomy column from provider+type ("esco_task",
"onet_knowledge", ...) to provider only, and dropped those names from
_expand_allowed_sources. A caller passing a granular source previously matched
the old labels; afterwards it fell through the map unchanged and matched
nothing, returning an empty result with no error. The six names that were
removed are restored as aliases folding onto their provider, so the old calls
work against the new data. Verified against th
[…]
