### 2026-09-14 1858142
ci: remove the TestPyPI preview package workflow
deploy-preview.yml built and uploaded a laiser-preview dev release to
TestPyPI on every PR push. It is not needed: the test job already exercises
the package, and Build runs `python -m build` plus `twine check` on every PR,
which covers the packaging path it was validating.

It was also the only failing check on #440, rejected by TestPyPI with
"400 File too large. Limit for project 'laiser-preview' is 100 MB". Removing
it clears that check, but not the cause: the same limit applies on PyPI, so a
v* tag will still be refused by deploy-prod until 
[…]
