### 2026-09-14 5521e5e
ci(release): publish release notes per tag instead of committing to main
The Changelog workflow committed a regenerated CHANGELOG.md straight to main
on every merge. main now has a ruleset requiring a pull request and passing
status checks, so that push is rejected unless the workflow's app is on the
bypass list — and a bypass would exempt it from the same protection that #437
asked for.

The job now runs on v* tags, generates notes for just that tag with
`git-cliff --current`, and creates or updates the matching GitHub Release. It
never writes to a branch, so it needs no bypass. deploy-prod.y
[…]
