### 2026-09-14 0330ad3
docs(api): correct extract_concepts threshold defaults and timing
The API reference is now generated from these docstrings, and two of them
described behaviour the code does not have.

similarity_thresholds was documented as defaulting to
{"skill": 0.20, "knowledge": 0.45, "task": 0.55}, and similarity_threshold as
defaulting to 0.20. When neither is passed, extract_and_align_core starts from
DEFAULT_SIMILARITY_THRESHOLDS in config.py, which is
{"skill": 0.60, "knowledge": 0.50, "task": 0.50}; a scalar
similarity_threshold replaces all three, and similarity_thresholds overrides
the types it na
[…]
