+++ laiser/skill_extractor_refactored.py
@@ -180,11 +180,12 @@ class SkillExtractorRefactored:
         top_k : int, optional
             Maximum number of aligned items to return per document (default: 25)
         similarity_threshold : float, optional
-            Global minimum similarity score applied to all types unless overridden
-            by similarity_thresholds. Defaults to 0.20 for backward compatibility.
+            One minimum similarity score applied to every type. When omitted, the
+            per-type defaults below apply.
         similarity_thresholds : dict, optional
-            Per-type thresholds. Keys: "skill", "knowledge", "task".
-            Defaults: {"skill": 0.20, "knowledge": 0.45, "task": 0.55}
+            Per-type minimums, keys "skill", "knowledge", "task". Applied on top of
+            similarity_threshold for the types given.
+            Defaults: {"skill": 0.60, "knowledge": 0.50, "task": 0.50}
         levels : bool
             Whether to extract skill levels
         batch_size : int
@@ -201,7 +202,7 @@ class SkillExtractorRefactored:
             "edges" contains ENABLES edges (Knowledge → Task per skill).
             If False (default), return a plain pd.DataFrame.
         timing : bool, optional
-            If True, print per-stage timing for extraction and alignment steps.
+            Accepted for compatibility with benchmark callers; currently has no effect.
         output_csv_path : str, optional
             If provided, write normalized results to this CSV path.
 
