### 2026-09-14 54ee128
fix(llm): run Hugging Face models on CPU without an API key
Constructing SkillExtractorRefactored with a Hugging Face model id and
use_gpu=False could not produce any output. Reproduced with
Qwen/Qwen2.5-0.5B-Instruct and no provider key set:

    WARNING: Transformer model initialization failed: CUDA is required but
    not available for bitsandbytes.
    AttributeError: 'NoneType' object has no attribute 'generate'

Two independent bugs:

- load_model_from_transformer always requested 8-bit bitsandbytes
  quantization, which requires CUDA, so the model never loaded on CPU.
  Quantization now
[…]
