+++ tests/test_llm_router.py
@@ -1,0 +1,135 @@
+"""Offline tests for local-model routing. No model weights are downloaded."""
+
+import pytest
+import torch
+
+from laiser.llm_models import hugging_face_llm, llm_router, model_loader
+from laiser.llm_models.llm_router import LLMRouter
+
+
+class _Encoded(dict):
+    def to(self, device):
+        return self
+
+
+class FakeTokenizer:
+    pad_token_id = 0
+    eos_token_id = 9
+
+    def __init__(self, chat_template=None):
+        self.chat_template = chat_template
+        self.seen_text = None
+
+    def apply_chat_template(self, messages, tokenize, add_generation_prompt):
+        return "CHAT:" + messages[0]["content"]
+
+    def __call__(self, text, return_tensors):
+        self.seen_text = text
+        return _Encoded(input_ids=torch.tensor([[1, 2, 3]]))
+
+    def decode(self, ids, skip_special_tokens):
+        return " ".join(str(i) for i in ids.tolist())
+
+
+class FakeModel:
+    device = "cpu"
+
+    def __init__(self):
+        self.generate_kwargs = None
+
+    def generate(self, **kwargs):
+        self.generate_kwargs = kwargs
+        # The prompt tokens (1, 2, 3) come back first, then the completion.
+        return torch.tensor([[1, 2, 3, 7, 8]])
+
+
+def _bare_router(**attrs):
+    """An LLMRouter with no model loading, so routing can be tested in isolation."""
+    router = LLMRouter.__new__(LLMRouter)
+    router.model_id = "some/model"
+    router.use_gpu = False
+    router.hf_token = router.api_key = router.backend = None
+    router.llm = router.model = router.tokenizer = router.nlp = None
+    for key, value in attrs.items():
+        setattr(router, key, value)
+    return router
+
+
+def test_transformers_model_is_routed_to_transformers_generate(monkeypatch):
+    calls = []
+    monkeypatch.setattr(llm_router, "llm_generate", lambda *args, **kwargs: calls.append(args) or "ok")
+    monkeypatch.setattr(llm_router, "llm_generate_vllm", lambda *args, **kwargs: pytest.fail("routed to vLLM"))
+
+    router = _bare_router(tokenizer=FakeTokenizer(), model=FakeModel())
+
+    assert router.generate("prompt") == "ok"
+    assert calls and calls[0][0] == "prompt"
+
+
+def test_vllm_engine_still_routes_to_vllm(monkeypatch):
+    monkeypatch.setattr(llm_router, "llm_generate", lambda *args, **kwargs: pytest.fail("routed to transformers"))
+    monkeypatch.setattr(llm_router, "llm_generate_vllm", lambda prompt, llm: f"vllm:{llm}")
+
+    assert _bare_router(llm="engine").generate("prompt") == "vllm:engine"
+
+
+def test_llm_generate_returns_only_the_completion():
+    model = FakeModel()
+
+    result = hugging_face_llm.llm_generate("p", FakeTokenizer(), model, "some/model", use_gpu=False)
+
+    assert result == "7 8"
+    assert model.generate_kwargs["max_new_tokens"] == hugging_face_llm.MAX_NEW_TOKENS
+
+
+def test_llm_generate_applies_chat_template_when_present():
+    tokenizer = FakeTokenizer(chat_template="{{ template }}")
+
+    hugging_face_llm.llm_generate("describe", tokenizer, FakeModel(), "some/model", use_gpu=False)
+
+    assert tokenizer.seen_text == "CHAT:describe"
+
+
+@pytest.fixture
+def recorded_loads(monkeypatch):
+    loads = []
+
[…]
