### 2026-09-14 d948e6a
docs: build the documentation site from this repository
The documentation site was generated from a separate docs repository, which
drifted from the code and has been deleted. Documentation now lives in docs/
here and is reviewed with the changes it describes.

- MkDocs with Material, as before, plus mkdocstrings so the API reference is
  generated from docstrings rather than maintained by hand.
- Pages: home, installation, quickstart, usage, model providers, taxonomies,
  the existing examples, cookbook notebooks, and API reference. Every option
  of extract_concepts, every output column, and
[…]
