### 2026-09-14 c314289
docs(readme): say what LAiSER does, fix formatting, add keyless and Colab paths
Responds to the README points in #438.

- About: rewritten to describe the actual input and output (a table of
  documents in, one row per taxonomy match out) and to name the taxonomies
  (ESCO, O*NET, UK Skills Classification, Open Skills Network) instead of
  general claims about AI and insight.
- Architecture: each stage's description sat on an indented line under its
  numbered item, which GitHub renders run into the title. Stages are now
  "**Stage** - description" on one line. Common runtime options had the s
[…]
