+++ README.md
@@ -5,48 +5,50 @@
 <div align="center">
 <img src="https://i.imgur.com/XznvjNi.png" width="70%"/>
 <h1>Leveraging ​Artificial ​Intelligence for ​Skill ​Extraction &​ Research (LAiSER)</h1>
+
+<a href="https://laiser-software.github.io/extract-module/"><b>Documentation</b></a> ·
+<a href="https://pypi.org/project/laiser/"><b>PyPI</b></a> ·
+<a href="https://github.com/LAiSER-Software/laiser-cookbook"><b>Cookbook notebooks</b></a>
 </div>
 
+LAiSER turns free text such as job postings and course syllabi into structured skills, knowledge and tasks, each matched to an entry in ESCO, O\*NET, the UK Skills Classification or the Open Skills Network.
+
 ### Contents
-LAiSER is a tool that helps learners, educators and employers share trusted and mutually intelligible information about skills​.
 
 - [About](#about)
 - [Architecture](#architecture)
 - [Requirements](#requirements)
 - [Setup and Installation](#setup-and-installation)
 - [Usage](#usage)
+- [Try it in Google Colab](#try-it-in-google-colab)
 - [Funding](#funding)
 - [Authors](#authors)
 - [Partners](#partners)
 
 ## About
 
-LAiSER is an innovative tool that harnesses the power of artificial intelligence to simplify the extraction and analysis of skills. It is designed for learners, educators, and employers who want to gain reliable insights into skill sets, ensuring that the information shared is both trusted and mutually intelligible across various sectors.
+LAiSER is a Python package for turning unstructured text about work and learning into structured, comparable skill data.
 
-By leveraging state-of-the-art AI models, LAiSER automates the process of identifying and classifying skills from diverse data sources. This not only saves time but also enhances accuracy, making it easier for users to discover emerging trends and in-demand skills.
+You give it a table of documents — job postings, course descriptions, syllabi. For each document, a language model extracts the skills it describes and, optionally, the knowledge and tasks behind them. LAiSER then matches every extracted phrase to its closest entries in established taxonomies using embedding similarity. The result is a table with one row per match: the phrase as written, the taxonomy entry it corresponds to, which taxonomy that entry comes from, and a similarity score.
 
-The tool emphasizes standardization and transparency, offering a common framework that bridges the communication gap between different stakeholders. With LAiSER, educators can better align their teaching methods with industry requirements, and employers can more effectively identify the competencies required for their teams. The result is a more efficient and strategic approach to skill development, benefiting the entire ecosystem.
+Because results point at shared taxonomy entries rather than free-text phrases, they can be counted and compared across sources — the skills employers ask for can be set against the skills a program teaches, or tracked across thousands of postings.
+
+The language model can be a hosted API (Gemini or OpenAI) or a model running on your own machine, including on CPU with no API key. Taxonomy alignment always runs locally against indexes bundled with the package.
 
 ## Architecture
 
-LAiSER uses a four-stage extraction and alignment pipeline:
+LAiSER runs four stages for every document:
 
-1. Extraction
-   Input text is normalized by input type and passed through prompt construction and LLM inference to produce raw concept candidates.
-2. Parsing and deduplication
-   Model output is parsed into structured concepts and filtered through exact and semantic deduplication.
-3. Taxonomy alignment
[…]
