### 2026-09-14 c689142
ci(release): publish release notes only after the package is published
Review of #441 pointed out that the release-notes workflow ran independently
of deploy-prod.yml on the same v* tags. A tag whose version did not match
laiser.__version__, whose wheel exceeded PyPI's per-file limit, or whose
upload failed would still get a public GitHub Release with notes and no
package behind it. With the wheel currently over the limit, that is exactly
what the next tag would have done.

Release notes are now a second job in deploy-prod.yml that needs the publish
job, so a Release is created only after the 
[…]
