+++ .github/workflows/changelog.yml
@@ -1,46 +1,0 @@
-name: Release Notes
-
-# Release notes are generated per tag and published on the GitHub Release,
-# rather than committed to CHANGELOG.md on main. main is protected (pull
-# request + required checks), so a workflow push there would be rejected, and
-# a bypass would weaken that protection for every other push.
-on:
-  push:
-    tags:
-      - "v*"
-
-concurrency:
-  group: release-notes-${{ github.ref }}
-  cancel-in-progress: false
-
-jobs:
-  release-notes:
-    name: Publish Release Notes
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - name: Generate notes for this tag
-        uses: orhun/git-cliff-action@v4
-        with:
-          config: cliff.toml
-          # --current: only the commits that belong to the tag being pushed.
-          args: --current --strip header
-        env:
-          OUTPUT: RELEASE_NOTES.md
-
-      - name: Create or update the GitHub Release
-        env:
-          GH_TOKEN: ${{ github.token }}
-          TAG: ${{ github.ref_name }}
-        run: |
-          if gh release view "$TAG" >/dev/null 2>&1; then
-            gh release edit "$TAG" --notes-file RELEASE_NOTES.md
-          else
-            gh release create "$TAG" --title "$TAG" --notes-file RELEASE_NOTES.md --verify-tag
-          fi
