### 2026-09-14 bc965b9
fix: address review findings on #442
Every point in Copilot's review was checked against the code; all held.

Code
- LLMRouter no longer swallows a local model load failure. The CPU branch
  caught the loader's exception and returned a router with no model; every
  generate() call then failed and extract_and_align_core skipped the row
  silently. Construction now raises LAiSERError, on CPU and on GPU when both
  vLLM and the transformers fallback fail. A vLLM failure alone still falls
  back to transformers.
- LLAMA_CPP_THREADS was read from the environment as a string and passed to
  llama.cp
[…]
