+++ laiser/skill_extractor_refactored.py
@@ -178,18 +178,20 @@ class SkillExtractorRefactored:
         input_type : str
             Type of input data
         top_k : int, optional
-            Maximum number of aligned items to return per document (default: 25)
+            Maximum number of aligned matches per concept type, per document
+            (default: 25). With skills, knowledge and tasks all requested, one
+            document can return up to three times this many rows.
         similarity_threshold : float, optional
             One minimum similarity score applied to every type. When omitted, the
             per-type defaults below apply.
         similarity_thresholds : dict, optional
-            Per-type minimums, keys "skill", "knowledge", "task". Applied on top of
-            similarity_threshold for the types given.
+            Per-type minimums, keys "skill", "knowledge", "task". Overrides
+            similarity_threshold for the types it names.
             Defaults: {"skill": 0.60, "knowledge": 0.50, "task": 0.50}
         levels : bool
-            Whether to extract skill levels
+            Accepted for compatibility; currently has no effect.
         batch_size : int
-            Batch size for processing
+            Accepted for compatibility; currently has no effect.
         warnings : bool
             Whether to show warnings
         concepts : list, optional
