### 2026-09-14 494938a
feat(llm): decode deterministically by default on every backend
Re-lands the decoding half of #426, which #428 reverted, ported onto the
current router and with the problems its review found fixed. Addresses the
non-determinism question in JOSS review #424.

Every backend previously inherited its provider's default decoding: vLLM,
OpenAI and Anthropic sample at temperature 1.0, llama.cpp used 0.2, and a
Transformers model used whatever its checkpoint's generation_config.json
said. Qwen2.5-0.5B-Instruct, for example, ships do_sample=true and
temperature=0.7, so the local CPU path sampled even t
[…]
