+++ tests/test_determinism.py
@@ -1,0 +1,458 @@
+"""Decoding determinism and alignment stability.
+
+Language models sample their output, so the same input can produce different
+generations across runs and across backends (JOSS review #424). LAiSER limits
+this in two ways, and these tests check both without a GPU, a model download or
+a paid API call:
+
+- Decoding: every backend defaults to greedy decoding and a fixed seed. Checked
+  against each entry point's defaults and against the parameters each backend
+  actually receives.
+- Alignment: generated phrases are matched to a fixed taxonomy by exact search,
+  so repeated alignment of the same phrases returns an identical result.
+"""
+
+import importlib
+import inspect
+from types import SimpleNamespace
+
+import pandas as pd
+import pytest
+import torch
+
+from laiser import config
+
+pytestmark = pytest.mark.determinism
+
+
+def _import_or_skip(module_path):
+    """Import an optional backend module, skipping when its dependency is absent.
+
+    pytest.importorskip only skips on ModuleNotFoundError, but ``from google import
+    genai`` without google-genai installed raises a plain ImportError.
+    """
+    try:
+        return importlib.import_module(module_path)
+    except ImportError as e:
+        pytest.skip(f"{module_path} unavailable in this environment: {e}")
+
+
+@pytest.fixture
+def reload_config(monkeypatch):
+    """Reload laiser.config under a controlled environment, then restore it.
+
+    laiser.config reads LAISER_TEMPERATURE and LAISER_TOP_P at import. Reloading
+    lets a test set or clear them without depending on the runner's environment;
+    teardown restores the variables and reloads again.
+    """
+    yield lambda: importlib.reload(config)
+    monkeypatch.undo()
+    importlib.reload(config)
+
+
+# ---------------------------------------------------------------------------
+# 1. Configuration
+# ---------------------------------------------------------------------------
+
+
+def test_shipped_defaults_are_greedy(monkeypatch, reload_config):
+    monkeypatch.delenv("LAISER_TEMPERATURE", raising=False)
+    monkeypatch.delenv("LAISER_TOP_P", raising=False)
+
+    fresh = reload_config()
+
+    assert fresh.DEFAULT_TEMPERATURE == 0.0
+    assert fresh.DEFAULT_TOP_P == 1.0
+    assert fresh.GENERATION_SEED == 42
+
+
+def test_environment_can_override_decoding_defaults(monkeypatch, reload_config):
+    monkeypatch.setenv("LAISER_TEMPERATURE", "0.7")
+    monkeypatch.setenv("LAISER_TOP_P", "0.9")
+
+    fresh = reload_config()
+
+    assert fresh.DEFAULT_TEMPERATURE == 0.7
+    assert fresh.DEFAULT_TOP_P == 0.9
+
+
+# ---------------------------------------------------------------------------
+# 2. Entry-point defaults
+# ---------------------------------------------------------------------------
+
+SEEDABLE_ENTRY_POINTS = [
+    ("laiser.llm_models.gemini", "gemini_generate"),
+    ("laiser.llm_models.gemini", "GeminiAPI"),
+    ("laiser.llm_models.hugging_face_llm", "llm_generate"),
+    ("laiser.llm_models.hugging_face_llm", "llm_generate_vllm"),
+    ("laiser.llm_models.llama_cpp_handler", "llama_cpp_chat"),
+    ("laiser.llm_models.llama_cpp_handler", "LlamaCppBackend"),
+    ("laiser.llm_models.llm_router", "LLMRouter"),
+    ("laiser.services", "SkillExtractionService"),
[…]
