### 2026-09-14 5b7d302
docs(paper): respond to JOSS review #424
Abbreviations. Every abbreviation is now spelled out at or before first use:
ESCO, LLM, AI, NER, FAISS, API, GPU, KSA, SCQF, MIT, GW and CCAS; spaCy is
described as a natural language processing library and EMSI as now
Lightcast. The ESCO and FAISS expansions had existed in earlier drafts but
were cut when the paper was shortened in be462ee.

Summary. Rewritten for a non-technical reader: it opens on the problem in
plain terms, explains what a skill taxonomy is before naming one, and says
who the software is for. Pipeline internals, the SCQF sentence, a
[…]
