+++ paper.md
@@ -29,7 +29,9 @@ bibliography: paper.bib
 
 # Summary
 
-LAiSER is an open-source Python framework for extracting, standardizing, and aligning skill information from unstructured text with established skill taxonomies. The system addresses the lack of standardized, machine-readable skill information needed to facilitate communication between learners, educators, and employers. LAiSER employs a two-stage pipeline combining large language models (LLMs) with semantic vector search using FAISS [@Johnson2019] for high-precision skill extraction and taxonomy alignment to standards such as ESCO [@ESCO2020]. The framework supports multiple computational backends, including GPU-accelerated inference, cloud APIs, and CPU-only environments, making it accessible across diverse research and operational contexts. LAiSER also supports extraction within the Knowledge, Skills, and Abilities (KSA) framework with automatic proficiency level classification using the Scottish Credit and Qualifications Framework (SCQF) [@SCQF2019]. The source code is available at [https://github.com/LAiSER-Software/extract-module](https://github.com/LAiSER-Software/extract-module) under the BSD 3-Clause License, installable via `pip install laiser`, or `pip install laiser[gpu]` for GPU inference.
+Job postings, course syllabi, and credential descriptions all describe skills, but each in its own words, so the same ability can be phrased many different ways. That makes it hard to count which skills employers ask for, or to check whether a course teaches what jobs require. Skill taxonomies address this by giving each skill one agreed name and identifier; the European Commission's European Skills, Competences, Qualifications and Occupations classification (ESCO) [@ESCO2020] is a widely used example.
+
+LAiSER is open-source software that reads such documents and links each skill they mention to its entry in a taxonomy. It uses a large language model (LLM), a kind of artificial intelligence (AI) system that reads and writes text, to find the skills a document describes, then matches each one to the closest taxonomy entry, so that differently worded mentions of the same skill are counted together. It is designed for researchers studying labor markets, curriculum designers comparing programs with employer demand, and organizations that issue digital credentials.
 
 # Statement of need
 
@@ -37,23 +39,29 @@ The contemporary labor market is characterized by rapid technological change, ev
 
 # State of the field
 
-Several open-source tools address parts of the skill extraction problem. SkillNER [@SkillNER2022] uses rule-based matching against the EMSI skills database via spaCy, offering fast lookups but limited ability to detect novel or context-dependent skills not present in its dictionary. Nesta's Skills Extractor Library [@Nesta2023] trains a spaCy NER model on labeled job advertisements and maps extracted phrases to ESCO or Lightcast taxonomies using sentence-transformer embeddings. While effective for UK job advertisements, its NER model requires domain-specific labeled training data and does not generalize well to other text types such as syllabi or credential descriptions. The esco-skill-extractor [@EscoExtractor2024] takes a simpler approach, embedding full input texts with a transformer and matching against ESCO entries by cosine similarity, but it lacks preprocessing, multi-model flexibility, and proficiency-level classification.
[…]
