### 2026-09-14 26a7b88
docs: document reproducible results and the decoding defaults
- New Reproducibility page: per-backend temperature and seed, why sampling is
  switched off explicitly for Transformers, how to change the defaults, a
  reference configuration, what is not guaranteed (hosted models change,
  different models extract different phrases, hardware and library versions),
  and what tests/test_determinism.py checks.
- README: temperature and seed in the runtime options table, and a
  "Reproducible results" section with the reference configuration, which the
  reply on #424 described but which was never 
[…]
