### 2026-09-15 25c1314
fix(gpu): install 8-bit dependencies with laiser[gpu]; correct provider docs
Addresses Copilot's review of #443.

- The gpu extra installed only vllm. When vLLM cannot load a model, LAiSER loads it with transformers using BitsAndBytesConfig(load_in_8bit=True) and device_map="auto", and transformers raises ImportError without bitsandbytes and accelerate, so the documented fallback could never run from a documented install. Both are added, with the minimum versions transformers itself checks for (bitsandbytes 0.43.1, accelerate 0.26.0).

- providers.md said a GPU model that fails to load falls b
[…]
