### 2026-09-15 c1fd5b7
fix: address review findings on #444
Code
- Gemini ignored LAISER_TOP_P: GenerateContentConfig was built without
  top_p, so the documented process-wide override reached every other
  backend but not Gemini. GeminiAPI and gemini_generate now take top_p
  (default DEFAULT_TOP_P) and send it; the Gemini request test asserts it.

Tests
- _import_or_skip skipped on any ImportError, so a broken import inside a
  backend module would have turned its tests into silent skips. It now
  skips only when ImportError.name names a genuinely optional package
  (google, vllm, llama_cpp) and re-raises everythi
[…]
