+++ paper.md
@@ -41,19 +41,19 @@ The contemporary labor market is characterized by rapid technological change, ev
 
 Several open-source tools address parts of the skill extraction problem. SkillNER [@SkillNER2022] uses rule-based matching against the EMSI (now Lightcast) skills database via spaCy, an open-source natural language processing library, offering fast lookups but limited ability to detect novel or context-dependent skills not present in its dictionary. Nesta's Skills Extractor Library [@Nesta2023] trains a spaCy named entity recognition (NER) model on labeled job advertisements and maps extracted phrases to ESCO or Lightcast taxonomies using sentence-transformer embeddings. While effective for UK job advertisements, its NER model requires domain-specific labeled training data and does not generalize well to other text types such as syllabi or credential descriptions. The esco-skill-extractor [@EscoExtractor2024] takes a simpler approach, embedding full input texts with a transformer and matching against ESCO entries by cosine similarity, but it lacks preprocessing, multi-model flexibility, and proficiency-level classification.
 
-LAiSER differs from these tools in several key respects. First, it uses LLMs for skill extraction rather than rule-based or NER-based approaches, enabling it to identify skills expressed in varied natural language without requiring labeled training data. Second, LAiSER provides domain-aware preprocessing that removes extraneous content (company branding, legal boilerplate, benefits sections) while preserving task-relevant skill language, an important step that existing tools omit. Third, the framework supports multiple LLM backends (vLLM, HuggingFace Transformers, llama.cpp, and the Gemini and OpenAI application programming interfaces, or APIs), allowing deployment on graphics processing unit (GPU) clusters, hosted APIs, or ordinary laptops. Finally, LAiSER uniquely supports Knowledge, Skills, and Abilities (KSA) decomposition with proficiency-level classification under the Scottish Credit and Qualifications Framework (SCQF) [@SCQF2019], a capability absent from all comparable tools. These design choices reflect a deliberate decision to build a new framework rather than extend existing tools, whose architectures are not designed to accommodate LLM-based extraction or multi-taxonomy proficiency alignment.
+LAiSER differs from these tools in several key respects. First, it uses LLMs for skill extraction rather than rule-based or NER-based approaches, enabling it to identify skills expressed in varied natural language without requiring labeled training data. Second, LAiSER's extraction prompts instruct the model to disregard extraneous content (company branding, legal boilerplate, benefits sections) while preserving task-relevant skill language, an important step that existing tools omit. Third, the framework supports multiple LLM backends (vLLM, HuggingFace Transformers, llama.cpp, and the Gemini and OpenAI application programming interfaces, or APIs), allowing deployment on graphics processing unit (GPU) clusters, hosted APIs, or ordinary laptops. Finally, LAiSER uniquely supports Knowledge, Skills, and Abilities (KSA) decomposition with proficiency-level classification under the Scottish Credit and Qualifications Framework (SCQF) [@SCQF2019], a capability absent from all comparable tools. These design choices reflect a deliberate decision to build a new framework rather than extend existing tools, whose architectures are not designed to accommodate LLM-based extraction or multi-taxonomy proficiency alignment.
 
 # Software design
 
 LAiSER's architecture is organized into three loosely coupled layers: data access, skill extraction, and taxonomy alignment. This separation was chosen to allow each component to evolve independently and to support diverse deployment scenarios without requiring changes to the overall pipeline.
 
[…]
