### 2026-09-15 05bc8bd
docs(paper): remove the unsupported SCQF proficiency claim
State of the field claimed LAiSER uniquely supports KSA decomposition with SCQF proficiency-level classification. The public API does not reach that capability: the KSA/SCQF prompt builders and parsers in services.py are never called, and extract_concepts(levels=...) has no effect. The sentence is removed, along with the related 'proficiency-level classification' and 'multi-taxonomy proficiency alignment' wording, so the paper only claims what the code does. Raised in Copilot's review of #444.

Co-Authored-By: Claude Opus 5 <noreply@an
[…]
